Small improvements and fixes
Added shortcode for SVG icons
Cosmetic changes in loader
Improved styles
