+++ rmcommon/js/cu-handler.js
@@ -1,4 +1,4 @@
-!function(t,e){"use strict";"function"==typeof define&&define.amd?define(["jquery"],e):"object"==typeof exports?module.exports=e(require("jquery")):t.bootbox=e(t.jQuery)}(this,function t(e,n){"use strict";function o(t){var e=m[f.locale];return e?e[t]:m.en[t]}function r(t,n,o){t.stopPropagation(),t.preventDefault();var r=e.isFunction(o)&&o.call(n,t)===!1;r||n.modal("hide")}function i(t){var e,n=0;for(e in t)n++;return n}function l(t,n){var o=0;e.each(t,function(t,e){n(t,e,o++)})}function a(t){var n,o;if("object"!=typeof t)throw new Error("Please supply an object of options");if(!t.message)throw new Error("Please specify a message");return t=e.extend({},f,t),t.buttons||(t.buttons={}),n=t.buttons,o=i(n),l(n,function(t,r,i){if(e.isFunction(r)&&(r=n[t]={callback:r}),"object"!==e.type(r))throw new Error("button with key "+t+" must be an object");r.label||(r.label=t),r.className||(2>=o&&i===o-1?r.className="btn-primary":r.className="btn-default")}),t}function s(t,e){var n=t.length,o={};if(1>n||n>2)throw new Error("Invalid argument length");return 2===n||"string"==typeof t[0]?(o[e[0]]=t[0],o[e[1]]=t[1]):o=t[0],o}function c(t,n,o){return e.extend(!0,{},t,s(n,o))}function u(t,e,n,o){var r={className:"bootbox-"+t,buttons:d.apply(null,e)};return p(c(r,o,n),e)}function d(){for(var t={},e=0,n=arguments.length;n>e;e++){var r=arguments[e],i=r.toLowerCase(),l=r.toUpperCase();t[i]={label:o(l)}}return t}function p(t,e){var o={};return l(e,function(t,e){o[e]=!0}),l(t.buttons,function(t){if(o[t]===n)throw new Error("button key "+t+" is not allowed (options are "+e.join("\n")+")")}),t}var h={dialog:"<div class='bootbox modal' tabindex='-1' role='dialog'><div class='modal-dialog'><div class='modal-content'><div class='modal-body'><div class='bootbox-body'></div></div></div></div></div>",header:"<div class='modal-header'><h4 class='modal-title'></h4></div>",footer:"<div class='modal-footer'></div>",closeButton:"<button type='button' class='bootbox-close-button close' data-dismiss='modal' aria-hidden='true'>&times;</button>",form:"<form class='bootbox-form'></form>",inputs:{text:"<input class='bootbox-input bootbox-input-text form-control' autocomplete=off type=text />",textarea:"<textarea class='bootbox-input bootbox-input-textarea form-control'></textarea>",email:"<input class='bootbox-input bootbox-input-email form-control' autocomplete='off' type='email' />",select:"<select class='bootbox-input bootbox-input-select form-control'></select>",checkbox:"<div class='checkbox'><label><input class='bootbox-input bootbox-input-checkbox' type='checkbox' /></label></div>",date:"<input class='bootbox-input bootbox-input-date form-control' autocomplete=off type='date' />",time:"<input class='bootbox-input bootbox-input-time form-control' autocomplete=off type='time' />",number:"<input class='bootbox-input bootbox-input-number form-control' autocomplete=off type='number' />",password:"<input class='bootbox-input bootbox-input-password form-control' autocomplete='off' type='password' />"}},f={locale:"en",backdrop:"static",animate:!0,className:null,closeButton:!0,show:!0,container:"body",id:""},b={};b.alert=function(){var t;if(t=u("alert",["ok"],["message","callback"],arguments),t.callback&&!e.isFunction(t.callback))throw new Error("alert requires callback property to be a function when provided");return t.buttons.ok.callback=t.onEscape=function(){return e.isFunction(t.callback)?t.callback.call(this):!0},b.dialog(t)},b.confirm=function(){var t;if(t=u("confirm",["cancel","confirm"],["message","callback"],arguments),t.buttons.cancel.callback=t.onEscape=function(){return t.callback.call(this,!1)},t.buttons.confirm.callback=function(){return t.callback.call(this,!0)},!e.isFunction(t.callback))throw new Error("confirm requires a callback");return b.dialog(t)},b.prompt=function(){var t,o,r,i,a,s,u;if(i=e(h.form),o={className:"bootbox-prompt",buttons:d("cancel","confirm"),value:"",inputType:"text"},t=p(c(o,arguments,["title","callback"]),["cancel","confirm"]),s=t.show===n?!0:t.show,t.message=i,t.buttons.cancel.callback=t.onEscape=function(){return t.callback.call(this,null)},t.buttons.confirm.callback=function(){var n;switch(t.inputType){case"text":case"textarea":case"email":case"select":case"date":case"time":case"number":case"password":n=a.val();break;case"checkbox":var o=a.find("input:checked");n=[],l(o,function(t,o){n.push(e(o).val())})}return t.callback.call(this,n)},t.show=!1,!t.title)throw new Error("prompt requires a title");if(!e.isFunction(t.callback))throw new Error("prompt requires a callback");if(!h.inputs[t.inputType])throw new Error("invalid prompt type");switch(a=e(h.inputs[t.inputType]),t.inputType){case"text":case"textarea":case"email":case"date":case"time":case"number":case"password":a.val(t.value);break;case"select":var f={};if(u=t.inputOptions||[],!e.isArray(u))throw new Error("Please pass an array of input options");if(!u.length)throw new Error("prompt with select requires options");l(u,function(t,o){var r=a;if(o.value===n||o.text===n)throw new Error("given options in wrong format");o.group&&(f[o.group]||(f[o.group]=e("<optgroup/>").attr("label",o.group)),r=f[o.group]),r.append("<option value='"+o.value+"'>"+o.text+"</option>")}),l(f,function(t,e){a.append(e)}),a.val(t.value);break;case"checkbox":var m=e.isArray(t.value)?t.value:[t.value];if(u=t.inputOptions||[],!u.length)throw new Error("prompt with checkbox requires options");if(!u[0].value||!u[0].text)throw new Error("given options in wrong format");a=e("<div/>"),l(u,function(n,o){var r=e(h.inputs[t.inputType]);r.find("input").attr("value",o.value),r.find("label").append(o.text),l(m,function(t,e){e===o.value&&r.find("input").prop("checked",!0)}),a.append(r)})}return t.placeholder&&a.attr("placeholder",t.placeholder),t.pattern&&a.attr("pattern",t.pattern),t.maxlength&&a.attr("maxlength",t.maxlength),i.append(a),i.on("submit",function(t){t.preventDefault(),t.stopPropagation(),r.find(".btn-primary").click()}),r=b.dialog(t),r.off("shown.bs.modal"),r.on("shown.bs.modal",function(){a.focus()}),s===!0&&r.modal("show"),r},b.dialog=function(t){t=a(t);var o=e(h.dialog),i=o.find(".modal-dialog"),s=o.find(".modal-body"),c=t.buttons,u="",d={onEscape:t.onEscape};if(e.fn.modal===n)throw new Error("$.fn.modal is not defined; please double check you have included the Bootstrap JavaScript library. See http://getbootstrap.com/javascript/ for more details.");if(l(c,function(t,e){u+="<button data-bb-handler='"+t+"' type='button' class='btn "+e.className+"'>"+e.label+"</button>",d[t]=e.callback}),s.find(".bootbox-body").html(t.message),t.id&&o.attr("id",t.id),t.animate===!0&&o.addClass("fade"),t.owner&&o.attr("data-owner",t.owner),t.className&&o.addClass(t.className),t.width&&(t.size=t.width),t.size&&i.addClass(t.size),t.color!=n&&""!=t.color&&i.addClass("modal-"+t.color),t.title&&s.before(h.header),t.closeButton){var p=e(h.closeButton);t.title?o.find(".modal-header").prepend(p):p.css("margin-top","-10px").prependTo(s)}return t.title&&o.find(".modal-title").html(t.icon?'<i class="'+t.icon+'"></i> '+t.title:t.title),u.length&&(s.after(h.footer),o.find(".modal-footer").html(u)),o.on("hidden.bs.modal",function(t){t.target===this&&o.remove()}),o.on("shown.bs.modal",function(){o.find(".btn-primary:first").focus()}),"static"!==t.backdrop&&o.on("click.dismiss.bs.modal",function(t){o.children(".modal-backdrop").length&&(t.currentTarget=o.children(".modal-backdrop").get(0)),t.target===t.currentTarget&&o.trigger("escape.close.bb")}),""!=t.id&&o.attr("id",t.id),o.on("escape.close.bb",function(t){d.onEscape&&r(t,o,d.onEscape)}),o.on("click",".modal-footer button",function(t){var n=e(this).data("bb-handler");r(t,o,d[n])}),o.on("click",".bootbox-close-button",function(t){r(t,o,d.onEscape)}),o.on("keyup",function(t){27===t.which&&o.trigger("escape.close.bb")}),e(t.container).append(o),o.modal({backdrop:t.backdrop?"static":!1,keyboard:!1,show:!1}),t.show&&o.modal("show"),o},b.setDefaults=function(){var t={};2===arguments.length?t[arguments[0]]=arguments[1]:t=arguments[0],e.extend(f,t)},b.hideAll=function(){return e(".bootbox").modal("hide"),b};var m={bg_BG:{OK:"Ок",CANCEL:"Отказ",CONFIRM:"Потвърждавам"},br:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Sim"},cs:{OK:"OK",CANCEL:"Zrušit",CONFIRM:"Potvrdit"},da:{OK:"OK",CANCEL:"Annuller",CONFIRM:"Accepter"},de:{OK:"OK",CANCEL:"Abbrechen",CONFIRM:"Akzeptieren"},el:{OK:"Εντάξει",CANCEL:"Ακύρωση",CONFIRM:"Επιβεβαίωση"},en:{OK:"OK",CANCEL:"Cancel",CONFIRM:"OK"},es:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Aceptar"},et:{OK:"OK",CANCEL:"Katkesta",CONFIRM:"OK"},fa:{OK:"قبول",CANCEL:"لغو",CONFIRM:"تایید"},fi:{OK:"OK",CANCEL:"Peruuta",CONFIRM:"OK"},fr:{OK:"OK",CANCEL:"Annuler",CONFIRM:"D'accord"},he:{OK:"אישור",CANCEL:"ביטול",CONFIRM:"אישור"},hu:{OK:"OK",CANCEL:"Mégsem",CONFIRM:"Megerősít"},hr:{OK:"OK",CANCEL:"Odustani",CONFIRM:"Potvrdi"},id:{OK:"OK",CANCEL:"Batal",CONFIRM:"OK"},it:{OK:"OK",CANCEL:"Annulla",CONFIRM:"Conferma"},ja:{OK:"OK",CANCEL:"キャンセル",CONFIRM:"確認"},lt:{OK:"Gerai",CANCEL:"Atšaukti",CONFIRM:"Patvirtinti"},lv:{OK:"Labi",CANCEL:"Atcelt",CONFIRM:"Apstiprināt"},nl:{OK:"OK",CANCEL:"Annuleren",CONFIRM:"Accepteren"},no:{OK:"OK",CANCEL:"Avbryt",CONFIRM:"OK"},pl:{OK:"OK",CANCEL:"Anuluj",CONFIRM:"Potwierdź"},pt:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Confirmar"},ru:{OK:"OK",CANCEL:"Отмена",CONFIRM:"Применить"},sq:{OK:"OK",CANCEL:"Anulo",CONFIRM:"Prano"},sv:{OK:"OK",CANCEL:"Avbryt",CONFIRM:"OK"},th:{OK:"ตกลง",CANCEL:"ยกเลิก",CONFIRM:"ยืนยัน"},tr:{OK:"Tamam",CANCEL:"İptal",CONFIRM:"Onayla"},zh_CN:{OK:"OK",CANCEL:"取消",CONFIRM:"确认"},zh_TW:{OK:"OK",CANCEL:"取消",CONFIRM:"確認"}};return b.addLocale=function(t,n){return e.each(["OK","CANCEL","CONFIRM"],function(t,e){if(!n[e])throw new Error("Please supply a translation for '"+e+"'")}),m[t]={OK:n.OK,CANCEL:n.CANCEL,CONFIRM:n.CONFIRM},b},b.removeLocale=function(t){return delete m[t],b},b.setLocale=function(t){return b.setDefaults("locale",t)},b.init=function(n){return t(n||e)},b});
-!function(t){t.fn.cuSpinner=function(e){var n=t.extend({icon:"svg-rmcommon-spinner-06",type:"spin",hide:"cu-icon",steps:12,speed:3},e);n.type="spin"!=n.type&&"pulse"!=n.type?"spin":n.type;"pulse"==n.type?"animation: cu-spin 1s infinite steps("+n.steps+")":"animation: cu-spin "+2/6*n.speed+"s infinite linear;";return this.each(function(){var e=t(this),o=cuHandler.getIcon(n.icon);o||(o=cuHandler.getIcon("svg-rmcommon-spinner-06"));var r=t(this).find(".cu-spinner");if(r.length>0)return t(r).remove(),t(this).find(".cu-spinner-hide").removeClass("cu-spinner-hide").show(),!0;var i=t("<span />",{"class":"cu-icon cu-spinner"});return i.load(o,function(){var o=t(e).find("."+n.hide);o.length>0?(o.addClass("cu-spinner-hide").hide(),t(o[0]).after(i)):t(e).prepend(i)}),!0})}}(jQuery);
-!function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery")):"function"==typeof define&&define.amd?define("pnotify",["jquery"],t):t(jQuery)}(function(t){var e,o,n={dir1:"down",dir2:"left",push:"bottom",spacing1:25,spacing2:25,context:t("body")},i=t(window),s=function(){o=t("body"),PNotify.prototype.options.stack.context=o,i=t(window),i.bind("resize",function(){e&&clearTimeout(e),e=setTimeout(function(){PNotify.positionAll(!0)},10)})};return PNotify=function(t){this.parseOptions(t),this.init()},t.extend(PNotify.prototype,{version:"2.1.0",options:{title:!1,title_escape:!1,text:!1,text_escape:!1,styling:"brighttheme",addclass:"",cornerclass:"",auto_display:!0,width:"300px",min_height:"16px",type:"notice",icon:!0,opacity:1,animation:"fade",animate_speed:"slow",position_animate_speed:500,shadow:!0,hide:!0,delay:8e3,mouse_reset:!0,remove:!0,insert_brs:!0,destroy:!0,stack:n},modules:{},runModules:function(t,e){var o;for(var n in this.modules)o="object"==typeof e&&n in e?e[n]:e,"function"==typeof this.modules[n][t]&&this.modules[n][t](this,"object"==typeof this.options[n]?this.options[n]:{},o)},state:"initializing",timer:null,styles:null,elem:null,container:null,title_container:null,text_container:null,animating:!1,timerHide:!1,init:function(){var e=this;if(this.modules={},t.extend(!0,this.modules,PNotify.prototype.modules),"object"==typeof this.options.styling?this.styles=this.options.styling:this.styles=PNotify.styling[this.options.styling],this.elem=t("<div />",{"class":"ui-pnotify "+this.options.addclass,css:{display:"none"},"aria-live":"assertive",mouseenter:function(t){if(e.options.mouse_reset&&"out"===e.animating){if(!e.timerHide)return;e.cancelRemove()}e.options.hide&&e.options.mouse_reset&&e.cancelRemove()},mouseleave:function(t){e.options.hide&&e.options.mouse_reset&&"out"!==e.animating&&e.queueRemove(),PNotify.positionAll()}}),this.container=t("<div />",{"class":"alert "+this.options.type,role:"alert"}).appendTo(this.elem),""!==this.options.cornerclass&&this.container.removeClass("ui-corner-all").addClass(this.options.cornerclass),this.options.shadow&&this.container.addClass("ui-pnotify-shadow"),this.options.icon!==!1)if("string"==typeof this.options.icon){var o=t("<div />",{"class":"ui-pnotify-icon"});cuHandler.loadIcon(this.options.icon,o),o.prependTo(this.container)}else t("<div />",{"class":"ui-pnotify-icon"}).append(t("<span />",{"class":this.options.icon===!0?"error"===this.options.type?this.styles.error_icon:"info"===this.options.type?this.styles.info_icon:"success"===this.options.type?this.styles.success_icon:this.styles.notice_icon:this.options.icon})).prependTo(this.container);return this.title_container=t("<h4 />",{"class":"ui-pnotify-title"}).appendTo(this.container),this.options.title===!1?this.title_container.hide():this.options.title_escape?this.title_container.text(this.options.title):this.title_container.html(this.options.title),this.text_container=t("<div />",{"class":"ui-pnotify-text"}).appendTo(this.container),this.options.text===!1?this.text_container.hide():this.options.text_escape?this.text_container.text(this.options.text):this.text_container.html(this.options.insert_brs?String(this.options.text).replace(/\n/g,"<br />"):this.options.text),"string"==typeof this.options.min_height&&this.container.css("min-height",this.options.min_height),"top"===this.options.stack.push?PNotify.notices=t.merge([this],PNotify.notices):PNotify.notices=t.merge(PNotify.notices,[this]),"top"===this.options.stack.push&&this.queuePosition(!1,1),this.options.stack.animation=!1,this.runModules("init"),this.options.auto_display&&this.open(),this},update:function(e){var o=this.options;if(this.parseOptions(o,e),this.options.cornerclass!==o.cornerclass&&this.container.removeClass("ui-corner-all "+o.cornerclass).addClass(this.options.cornerclass),this.options.shadow!==o.shadow&&(this.options.shadow?this.container.addClass("ui-pnotify-shadow"):this.container.removeClass("ui-pnotify-shadow")),this.options.addclass===!1?this.elem.removeClass(o.addclass):this.options.addclass!==o.addclass&&this.elem.removeClass(o.addclass).addClass(this.options.addclass),this.options.title===!1?this.title_container.slideUp("fast"):this.options.title!==o.title&&(this.options.title_escape?this.title_container.text(this.options.title):this.title_container.html(this.options.title),o.title===!1&&this.title_container.slideDown(200)),this.options.text===!1?this.text_container.slideUp("fast"):this.options.text!==o.text&&(this.options.text_escape?this.text_container.text(this.options.text):this.text_container.html(this.options.insert_brs?String(this.options.text).replace(/\n/g,"<br />"):this.options.text),o.text===!1&&this.text_container.slideDown(200)),this.options.type!==o.type&&this.container.removeClass(this.styles.error+" "+this.styles.notice+" "+this.styles.success+" "+this.styles.info).addClass(this.options.type),(this.options.icon!==o.icon||this.options.icon===!0&&this.options.type!==o.type)&&(this.container.find("div.ui-pnotify-icon").remove(),this.options.icon!==!1))if("string"==typeof this.options.icon){var n=t("<div />",{"class":"ui-pnotify-icon"});cuHandler.loadIcon(this.options.icon,n),n.prependTo(this.container)}else t("<div />",{"class":"ui-pnotify-icon"}).append(t("<span />",{"class":this.options.icon===!0?"error"===this.options.type?this.styles.error_icon:"info"===this.options.type?this.styles.info_icon:"success"===this.options.type?this.styles.success_icon:this.styles.notice_icon:this.options.icon})).prependTo(this.container);return this.options.min_height!==o.min_height&&this.container.animate({minHeight:this.options.min_height}),this.options.opacity!==o.opacity&&this.elem.fadeTo(this.options.animate_speed,this.options.opacity),this.options.hide?o.hide||this.queueRemove():this.cancelRemove(),this.queuePosition(!0),this.runModules("update",o),this},open:function(){this.state="opening",this.runModules("beforeOpen");var t=this;return this.elem.parent().length||this.elem.appendTo(this.options.stack.context?this.options.stack.context:o),"top"!==this.options.stack.push&&this.position(!0),"fade"===this.options.animation||"fade"===this.options.animation.effect_in?this.elem.show().fadeTo(0,0).hide():1!==this.options.opacity&&this.elem.show().fadeTo(0,this.options.opacity).hide(),this.animateIn(function(){t.queuePosition(!0),t.options.hide&&t.queueRemove(),t.state="open",t.runModules("afterOpen")}),this},remove:function(e){this.state="closing",this.timerHide=!!e,this.runModules("beforeClose");var o=this;return this.timer&&(window.clearTimeout(this.timer),this.timer=null),this.animateOut(function(){if(o.state="closed",o.runModules("afterClose"),o.queuePosition(!0),o.options.remove&&o.elem.detach(),o.runModules("beforeDestroy"),o.options.destroy&&null!==PNotify.notices){var e=t.inArray(o,PNotify.notices);-1!==e&&PNotify.notices.splice(e,1)}o.runModules("afterDestroy")}),this},get:function(){return this.elem},parseOptions:function(e,o){this.options=t.extend(!0,{},PNotify.prototype.options),this.options.stack=PNotify.prototype.options.stack;for(var n,i=[e,o],s=0;s<i.length&&(n=i[s],"undefined"!=typeof n);s++)if("object"!=typeof n)this.options.text=n;else for(var r in n)this.modules[r]?t.extend(!0,this.options[r],n[r]):this.options[r]=n[r]},animateIn:function(t){this.animating="in";var e;e="undefined"!=typeof this.options.animation.effect_in?this.options.animation.effect_in:this.options.animation,"none"===e?(this.elem.show(),t()):"show"===e?this.elem.show(this.options.animate_speed,t):"fade"===e?this.elem.show().fadeTo(this.options.animate_speed,this.options.opacity,t):"slide"===e?this.elem.slideDown(this.options.animate_speed,t):"function"==typeof e?e("in",t,this.elem):this.elem.show(e,"object"==typeof this.options.animation.options_in?this.options.animation.options_in:{},this.options.animate_speed,t),this.elem.parent().hasClass("ui-effects-wrapper")&&this.elem.parent().css({position:"fixed",overflow:"visible"}),"slide"!==e&&this.elem.css("overflow","visible"),this.container.css("overflow","hidden")},animateOut:function(t){this.animating="out";var e;e="undefined"!=typeof this.options.animation.effect_out?this.options.animation.effect_out:this.options.animation,"none"===e?(this.elem.hide(),t()):"show"===e?this.elem.hide(this.options.animate_speed,t):"fade"===e?this.elem.fadeOut(this.options.animate_speed,t):"slide"===e?this.elem.slideUp(this.options.animate_speed,t):"function"==typeof e?e("out",t,this.elem):this.elem.hide(e,"object"==typeof this.options.animation.options_out?this.options.animation.options_out:{},this.options.animate_speed,t),this.elem.parent().hasClass("ui-effects-wrapper")&&this.elem.parent().css({position:"fixed",overflow:"visible"}),"slide"!==e&&this.elem.css("overflow","visible"),this.container.css("overflow","hidden")},position:function(t){var e=this.options.stack,n=this.elem;if(n.parent().hasClass("ui-effects-wrapper")&&(n=this.elem.css({left:"0",top:"0",right:"0",bottom:"0"}).parent()),"undefined"==typeof e.context&&(e.context=o),e){"number"!=typeof e.nextpos1&&(e.nextpos1=e.firstpos1),"number"!=typeof e.nextpos2&&(e.nextpos2=e.firstpos2),"number"!=typeof e.addpos2&&(e.addpos2=0);var s="none"===n.css("display");if(!s||t){var r,a,l,c={};switch(e.dir1){case"down":l="top";break;case"up":l="bottom";break;case"left":l="right";break;case"right":l="left"}r=parseInt(n.css(l).replace(/(?:\..*|[^0-9.])/g,"")),isNaN(r)&&(r=0),"undefined"!=typeof e.firstpos1||s||(e.firstpos1=r,e.nextpos1=e.firstpos1);var p;switch(e.dir2){case"down":p="top";break;case"up":p="bottom";break;case"left":p="right";break;case"right":p="left"}if(a=parseInt(n.css(p).replace(/(?:\..*|[^0-9.])/g,"")),isNaN(a)&&(a=0),"undefined"!=typeof e.firstpos2||s||(e.firstpos2=a,e.nextpos2=e.firstpos2),("down"===e.dir1&&e.nextpos1+n.height()>(e.context.is(o)?i.height():e.context.prop("scrollHeight"))||"up"===e.dir1&&e.nextpos1+n.height()>(e.context.is(o)?i.height():e.context.prop("scrollHeight"))||"left"===e.dir1&&e.nextpos1+n.width()>(e.context.is(o)?i.width():e.context.prop("scrollWidth"))||"right"===e.dir1&&e.nextpos1+n.width()>(e.context.is(o)?i.width():e.context.prop("scrollWidth")))&&(e.nextpos1=e.firstpos1,e.nextpos2+=e.addpos2+("undefined"==typeof e.spacing2?25:e.spacing2),e.addpos2=0),e.animation&&e.nextpos2<a)switch(e.dir2){case"down":c.top=e.nextpos2+"px";break;case"up":c.bottom=e.nextpos2+"px";break;case"left":c.right=e.nextpos2+"px";break;case"right":c.left=e.nextpos2+"px"}else"number"==typeof e.nextpos2&&n.css(p,e.nextpos2+"px");switch(e.dir2){case"down":case"up":n.outerHeight(!0)>e.addpos2&&(e.addpos2=n.height());break;case"left":case"right":n.outerWidth(!0)>e.addpos2&&(e.addpos2=n.width())}if("number"==typeof e.nextpos1)if(e.animation&&(r>e.nextpos1||c.top||c.bottom||c.right||c.left))switch(e.dir1){case"down":c.top=e.nextpos1+"px";break;case"up":c.bottom=e.nextpos1+"px";break;case"left":c.right=e.nextpos1+"px";break;case"right":c.left=e.nextpos1+"px"}else n.css(l,e.nextpos1+"px");switch((c.top||c.bottom||c.right||c.left)&&n.animate(c,{duration:this.options.position_animate_speed,queue:!1}),e.dir1){case"down":case"up":e.nextpos1+=n.height()+("undefined"==typeof e.spacing1?25:e.spacing1);break;case"left":case"right":e.nextpos1+=n.width()+("undefined"==typeof e.spacing1?25:e.spacing1)}}return this}},queuePosition:function(t,o){return e&&clearTimeout(e),o||(o=10),e=setTimeout(function(){PNotify.positionAll(t)},o),this},cancelRemove:function(){return this.timer&&window.clearTimeout(this.timer),"closing"===this.state&&(this.elem.stop(!0),this.state="open",this.animating="in",this.elem.css("height","auto").animate({width:this.options.width,opacity:this.options.opacity},"fast")),this},queueRemove:function(){var t=this;return this.cancelRemove(),this.timer=window.setTimeout(function(){t.remove(!0)},isNaN(this.options.delay)?0:this.options.delay),this}}),t.extend(PNotify,{notices:[],removeAll:function(){t.each(PNotify.notices,function(){this.remove&&this.remove(!1)})},positionAll:function(o){if(e&&clearTimeout(e),e=null,PNotify.notices&&PNotify.notices.length)t.each(PNotify.notices,function(){var t=this.options.stack;t&&(t.nextpos1=t.firstpos1,t.nextpos2=t.firstpos2,t.addpos2=0,t.animation=o)}),t.each(PNotify.notices,function(){this.position()});else{var n=PNotify.prototype.options.stack;n&&(delete n.nextpos1,delete n.nextpos2)}},styling:{brighttheme:{container:"brighttheme",notice:"brighttheme-notice",notice_icon:"brighttheme-icon-notice",info:"brighttheme-info",info_icon:"brighttheme-icon-info",success:"brighttheme-success",success_icon:"brighttheme-icon-success",error:"brighttheme-error",error_icon:"brighttheme-icon-error"},jqueryui:{container:"ui-widget ui-widget-content ui-corner-all",notice:"ui-state-highlight",notice_icon:"ui-icon ui-icon-info",info:"",info_icon:"ui-icon ui-icon-info",success:"ui-state-default",success_icon:"ui-icon ui-icon-circle-check",error:"ui-state-error",error_icon:"ui-icon ui-icon-alert"},bootstrap2:{container:"alert",notice:"",notice_icon:"icon-exclamation-sign",info:"alert-info",info_icon:"icon-info-sign",success:"alert-success",success_icon:"icon-ok-sign",error:"alert-error",error_icon:"icon-warning-sign"},bootstrap3:{container:"alert",notice:"alert-warning",notice_icon:"glyphicon glyphicon-exclamation-sign",info:"alert-info",info_icon:"glyphicon glyphicon-info-sign",success:"alert-success",success_icon:"glyphicon glyphicon-ok-sign",error:"alert-danger",error_icon:"glyphicon glyphicon-warning-sign"}}}),PNotify.styling.fontawesome=t.extend({},PNotify.styling.bootstrap3),t.extend(PNotify.styling.fontawesome,{notice_icon:"fa fa-exclamation-circle",info_icon:"fa fa-info",success_icon:"fa fa-check",error_icon:"fa fa-warning"}),document.body?s():t(s),PNotify}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.buttons",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,e){e.prototype.options.buttons={closer:!0,closer_hover:!0,sticker:!0,sticker_hover:!0,show_on_nonblock:!1,labels:{close:"Close",stick:"Stick"}},e.prototype.modules.buttons={myOptions:null,closer:null,sticker:null,init:function(e,o){var n=this;this.myOptions=o,e.elem.on({mouseenter:function(t){!n.myOptions.sticker||e.options.nonblock&&e.options.nonblock.nonblock&&!n.myOptions.show_on_nonblock||n.sticker.trigger("pnotify_icon").css("visibility","visible"),!n.myOptions.closer||e.options.nonblock&&e.options.nonblock.nonblock&&!n.myOptions.show_on_nonblock||n.closer.css("visibility","visible")},mouseleave:function(t){n.myOptions.sticker_hover&&n.sticker.css("visibility","hidden"),n.myOptions.closer_hover&&n.closer.css("visibility","hidden")}}),this.sticker=t("<div />",{"class":"ui-pnotify-sticker",css:{cursor:"pointer",visibility:o.sticker_hover?"hidden":"visible"},click:function(){e.options.hide=!e.options.hide,e.options.hide?e.queueRemove():e.cancelRemove(),t(this).trigger("pnotify_icon")}}).bind("pnotify_icon",function(){t(this).children().removeClass(e.styles.pin_up+" "+e.styles.pin_down).addClass(e.options.hide?e.styles.pin_up:e.styles.pin_down),t(this).children().load(cuHandler.url("/modules/rmcommon/icons/"+(e.options.hide?"pause.svg":"play.svg")))}).append(t("<span />",{"class":e.styles.pin_up,title:o.labels.stick}).load(cuHandler.url("/modules/rmcommon/icons/pause.svg"))).prependTo(e.container),(!o.sticker||e.options.nonblock&&e.options.nonblock.nonblock&&!o.show_on_nonblock)&&this.sticker.css("display","none"),this.closer=t("<div />",{"class":"ui-pnotify-closer",css:{cursor:"pointer",visibility:o.closer_hover?"hidden":"visible"},click:function(){e.remove(!1),n.sticker.css("visibility","hidden"),n.closer.css("visibility","hidden")}}).append(t("<span />",{"class":e.styles.closer,title:o.labels.close}).load(cuHandler.url("/modules/rmcommon/icons/close.svg"))).prependTo(e.container),(!o.closer||e.options.nonblock&&e.options.nonblock.nonblock&&!o.show_on_nonblock)&&this.closer.css("display","none")},update:function(t,e){this.myOptions=e,!e.closer||t.options.nonblock&&t.options.nonblock.nonblock&&!e.show_on_nonblock?this.closer.css("display","none"):e.closer&&this.closer.css("display","block"),!e.sticker||t.options.nonblock&&t.options.nonblock.nonblock&&!e.show_on_nonblock?this.sticker.css("display","none"):e.sticker&&this.sticker.css("display","block"),this.sticker.trigger("pnotify_icon"),e.sticker_hover?this.sticker.css("visibility","hidden"):t.options.nonblock&&t.options.nonblock.nonblock&&!e.show_on_nonblock||this.sticker.css("visibility","visible"),e.closer_hover?this.closer.css("visibility","hidden"):t.options.nonblock&&t.options.nonblock.nonblock&&!e.show_on_nonblock||this.closer.css("visibility","visible")}},t.extend(e.styling.brighttheme,{closer:"cu-icon",pin_up:"cu-icon",pin_down:"cu-icon"}),t.extend(e.styling.jqueryui,{closer:"ui-icon ui-icon-close",pin_up:"ui-icon ui-icon-pin-w",pin_down:"ui-icon ui-icon-pin-s"}),t.extend(e.styling.bootstrap2,{closer:"icon-remove",pin_up:"icon-pause",pin_down:"icon-play"}),t.extend(e.styling.bootstrap3,{closer:"glyphicon glyphicon-remove",pin_up:"glyphicon glyphicon-pause",pin_down:"glyphicon glyphicon-play"}),t.extend(e.styling.fontawesome,{closer:"fa fa-times",pin_up:"fa fa-pause",pin_down:"fa fa-play"})}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.callbacks",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,e){var o=e.prototype.init,n=e.prototype.open,i=e.prototype.remove;e.prototype.init=function(){this.options.before_init&&this.options.before_init(this.options),o.apply(this,arguments),this.options.after_init&&this.options.after_init(this)},e.prototype.open=function(){var t;this.options.before_open&&(t=this.options.before_open(this)),t!==!1&&(n.apply(this,arguments),this.options.after_open&&this.options.after_open(this))},e.prototype.remove=function(t){var e;this.options.before_close&&(e=this.options.before_close(this,t)),e!==!1&&(i.apply(this,arguments),this.options.after_close&&this.options.after_close(this,t))}}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.confirm",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,e){e.prototype.options.confirm={confirm:!1,prompt:!1,prompt_class:"",prompt_default:"",prompt_multi_line:!1,align:"right",buttons:[{text:"Ok",addClass:"",promptTrigger:!0,click:function(t,e){t.remove(),t.get().trigger("pnotify.confirm",[t,e])}},{text:"Cancel",addClass:"",click:function(t){t.remove(),t.get().trigger("pnotify.cancel",t)}}]},e.prototype.modules.confirm={container:null,prompt:null,init:function(e,o){this.container=t('<div style="margin-top:5px;clear:both;" />').css("text-align",o.align).appendTo(e.container),o.confirm||o.prompt?this.makeDialog(e,o):this.container.hide()},update:function(t,e){e.confirm?(this.makeDialog(t,e),this.container.show()):this.container.hide().empty()},afterOpen:function(t,e){e.prompt&&this.prompt.focus()},makeDialog:function(e,o){var n,i,s=!1,r=this;this.container.empty(),o.prompt&&(this.prompt=t("<"+(o.prompt_multi_line?'textarea rows="5"':'input type="text"')+' style="margin-bottom:5px;clear:both;" />').addClass(e.styles.input+" "+o.prompt_class).val(o.prompt_default).appendTo(this.container));for(var a in o.buttons)n=o.buttons[a],s?this.container.append(" "):s=!0,i=t('<button type="button" />').addClass(e.styles.btn+" "+n.addClass).text(n.text).appendTo(this.container).on("click",function(t){return function(){"function"==typeof t.click&&t.click(e,o.prompt?r.prompt.val():null)}}(n)),o.prompt&&!o.prompt_multi_line&&n.promptTrigger&&this.prompt.keypress(function(t){return function(e){13==e.keyCode&&t.click()}}(i)),e.styles.text&&i.wrapInner('<span class="'+e.styles.text+'"></span>'),e.styles.btnhover&&i.hover(function(t){return function(){t.addClass(e.styles.btnhover)}}(i),function(t){return function(){t.removeClass(e.styles.btnhover)}}(i)),e.styles.btnactive&&i.on("mousedown",function(t){return function(){t.addClass(e.styles.btnactive)}}(i)).on("mouseup",function(t){return function(){t.removeClass(e.styles.btnactive)}}(i)),e.styles.btnfocus&&i.on("focus",function(t){return function(){t.addClass(e.styles.btnfocus)}}(i)).on("blur",function(t){return function(){t.removeClass(e.styles.btnfocus)}}(i))}},t.extend(e.styling.jqueryui,{btn:"ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only",btnhover:"ui-state-hover",btnactive:"ui-state-active",btnfocus:"ui-state-focus",input:"",text:"ui-button-text"}),t.extend(e.styling.bootstrap2,{btn:"btn",input:""}),t.extend(e.styling.bootstrap3,{btn:"btn btn-default",input:"form-control"}),t.extend(e.styling.fontawesome,{btn:"btn btn-default",input:"form-control"})}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.desktop",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,e){var o,n=function(t,e){return(n="Notification"in window?function(t,e){return new Notification(t,e)}:"mozNotification"in navigator?function(t,e){return navigator.mozNotification.createNotification(t,e.body,e.icon).show()}:"webkitNotifications"in window?function(t,e){return window.webkitNotifications.createNotification(e.icon,t,e.body)}:function(t,e){return null})(t,e)};e.prototype.options.desktop={desktop:!1,fallback:!0,icon:null,tag:null},e.prototype.modules.desktop={tag:null,icon:null,genNotice:function(t,e){null===e.icon?this.icon="http://sciactive.com/pnotify/includes/desktop/"+t.options.type+".png":e.icon===!1?this.icon=null:this.icon=e.icon,(null===this.tag||null!==e.tag)&&(this.tag=null===e.tag?"PNotify-"+Math.round(1e6*Math.random()):e.tag),t.desktop=n(t.options.title,{icon:this.icon,body:t.options.text,tag:this.tag}),!("close"in t.desktop)&&"cancel"in t.desktop&&(t.desktop.close=function(){t.desktop.cancel()}),t.desktop.onclick=function(){t.elem.trigger("click")},t.desktop.onclose=function(){"closing"!==t.state&&"closed"!==t.state&&t.remove()}},init:function(t,n){return n.desktop?(o=e.desktop.checkPermission(),0!==o?void(n.fallback||(t.options.auto_display=!1)):void this.genNotice(t,n)):void 0},update:function(t,e,n){0!==o&&e.fallback||!e.desktop||this.genNotice(t,e)},beforeOpen:function(t,e){0!==o&&e.fallback||!e.desktop||t.elem.css({left:"-10000px",display:"none"})},afterOpen:function(t,e){0!==o&&e.fallback||!e.desktop||(t.elem.css({left:"-10000px",display:"none"}),"show"in t.desktop&&t.desktop.show())},beforeClose:function(t,e){0!==o&&e.fallback||!e.desktop||t.elem.css({left:"-10000px",display:"none"})},afterClose:function(t,e){0!==o&&e.fallback||!e.desktop||(t.elem.css({left:"-10000px",display:"none"}),"close"in t.desktop&&t.desktop.close())}},e.desktop={permission:function(){"undefined"!=typeof Notification&&"requestPermission"in Notification?Notification.requestPermission():"webkitNotifications"in window&&window.webkitNotifications.requestPermission()},checkPermission:function(){return"undefined"!=typeof Notification&&"permission"in Notification?"granted"===Notification.permission?0:1:"webkitNotifications"in window&&0==window.webkitNotifications.checkPermission()?0:1}},o=e.desktop.checkPermission()}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.nonblock",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,e){var o,n=/^on/,i=/^(dbl)?click$|^mouse(move|down|up|over|out|enter|leave)$|^contextmenu$/,s=/^(focus|blur|select|change|reset)$|^key(press|down|up)$/,r=/^(scroll|resize|(un)?load|abort|error)$/,a=function(e,o){var a;if(e=e.toLowerCase(),document.createEvent&&this.dispatchEvent){if(e=e.replace(n,""),e.match(i)?(t(this).offset(),a=document.createEvent("MouseEvents"),a.initMouseEvent(e,o.bubbles,o.cancelable,o.view,o.detail,o.screenX,o.screenY,o.clientX,o.clientY,o.ctrlKey,o.altKey,o.shiftKey,o.metaKey,o.button,o.relatedTarget)):e.match(s)?(a=document.createEvent("UIEvents"),a.initUIEvent(e,o.bubbles,o.cancelable,o.view,o.detail)):e.match(r)&&(a=document.createEvent("HTMLEvents"),a.initEvent(e,o.bubbles,o.cancelable)),!a)return;this.dispatchEvent(a)}else e.match(n)||(e="on"+e),a=document.createEventObject(o),this.fireEvent(e,a)},l=function(e,n,i){e.elem.css("display","none");var s=document.elementFromPoint(n.clientX,n.clientY);e.elem.css("display","block");var r=t(s),l=r.css("cursor");"auto"===l&&"A"===s.tagName&&(l="pointer"),e.elem.css("cursor","auto"!==l?l:"default"),o&&o.get(0)==s||(o&&(a.call(o.get(0),"mouseleave",n.originalEvent),a.call(o.get(0),"mouseout",n.originalEvent)),a.call(s,"mouseenter",n.originalEvent),a.call(s,"mouseover",n.originalEvent)),a.call(s,i,n.originalEvent),o=r};e.prototype.options.nonblock={nonblock:!1,nonblock_opacity:.2},e.prototype.modules.nonblock={myOptions:null,init:function(t,e){var n=this;this.myOptions=e,t.elem.on({mouseenter:function(e){n.myOptions.nonblock&&e.stopPropagation(),n.myOptions.nonblock&&t.elem.stop().animate({opacity:n.myOptions.nonblock_opacity},"fast")},mouseleave:function(e){n.myOptions.nonblock&&e.stopPropagation(),o=null,t.elem.css("cursor","auto"),n.myOptions.nonblock&&"out"!==t.animating&&t.elem.stop().animate({opacity:t.options.opacity},"fast")},mouseover:function(t){n.myOptions.nonblock&&t.stopPropagation()},mouseout:function(t){n.myOptions.nonblock&&t.stopPropagation()},mousemove:function(e){n.myOptions.nonblock&&(e.stopPropagation(),l(t,e,"onmousemove"))},mousedown:function(e){n.myOptions.nonblock&&(e.stopPropagation(),e.preventDefault(),l(t,e,"onmousedown"))},mouseup:function(e){n.myOptions.nonblock&&(e.stopPropagation(),e.preventDefault(),l(t,e,"onmouseup"))},click:function(e){n.myOptions.nonblock&&(e.stopPropagation(),l(t,e,"onclick"))},dblclick:function(e){n.myOptions.nonblock&&(e.stopPropagation(),l(t,e,"ondblclick"))}})},update:function(t,e){this.myOptions=e}}});
-var cuHandler={ismobile:/android|webos|iphone|ipad|ipod|blackberry|iemobile|opera mini/i.test(navigator.userAgent.toLowerCase()),currentResponse:!1,loadRemoteDialog:function(launcher){var url=void 0!=$(launcher).attr("href")&&"#"!=$(launcher).attr("href")?$(launcher).attr("href"):$(launcher).data("url"),handler=$(launcher).data("handler"),window_id=$(launcher).data("window-id"),params={CUTOKEN_REQUEST:$("#cu-token").val()};return 0==params?!1:(cuHandler.showLoader(),$.get(url,params,function(response){return cuHandler.retrieveAjax(response)?(void 0==handler&&""!=handler?(cuHandler.closeLoader(),bootbox.dialog({message:response.content,title:response.message,icon:void 0!=response.icon?response.icon:"",width:void 0!=response.width?response.width:"",id:void 0!=response.id?response.id:window_id,animate:!1,color:void 0!=response.color?response.color:"",closeButton:void 0!=response.closeButton?response.closeButton:!0})):(cuHandler.closeLoader(),eval(handler+"(response, launcher);")),$(".cu-data-table").each(function(){cuHandler.createDataTable($(this))}),cuHandler.checkAjaxAction(response),!1):!1},"json"),!1)},submitAjaxForm:function(t){if(!$(t).valid())return!1;cuHandler.showLoader();var e=t.serialize();e+="&CUTOKEN_REQUEST="+$("#cu-token").val();var o=t.attr("action"),n=t.attr("method");return"post"==n?$.post(o,e,cuHandler.retrieveAjax,"json"):$.get(o,e,cuHandler.retrieveAjax,"json"),!1},requestAjax:function(t,e){return $(t).cuSpinner({icon:"svg-rmcommon-spinner-14"}),void 0==e?!1:void 0==e.url?!1:(void 0==e.parameters&&(e.parameters={}),e.parameters.CUTOKEN_REQUEST=$("#cu-token").val(),void(void 0==e.method||"post"==e.method?$.post(e.url,e.parameters,cuHandler.retrieveAjax,"json"):$.get(e.url,e.parameters,cuHandler.retrieveAjax,"json")))},isDisabled:function(t){return $(t).hasClass("disabled")||void 0!=$(t).attr("disabled")?!0:!1},retrieveAjax:function(t){return this.currentResponse=t,"error"==t.type&&(void 0!=t.modal_message?bootbox.alert({message:t.message}):void 0!=t.notify&&void 0!=t.message?(cuHandler.notify({title:void 0==t.notify.title?"":t.notify.title,type:void 0==t.notify.type?"alert-info":t.notify.type,icon:void 0==t.notify.icon?"svg-rmcommon-info-solid":t.notify.icon,text:t.message}),t.notify=void 0):alert(t.message)),void 0!=t.token&&""!=t.token&&$("#cu-token").val(t.token),cuHandler.closeLoader(),cuHandler.checkAjaxAction(t),"error"==t.type?!1:!0},checkAjaxAction:function(data){return void 0!=data.showMessage&&alert(data.message),void 0!=data.notify&&void 0!=data.message&&this.notify({title:void 0==data.notify.title?null:data.notify.title,type:void 0==data.notify.type?"alert-info":data.notify.type,icon:void 0==data.notify.icon?"svg-rmcommon-info-solid":data.notify.icon,text:data.message}),void 0!=data.closeWindow&&$(data.closeWindow).modal("hide"),void 0!=data.runHandler&&eval(data.runHandler+"(data)"),void 0!=data["goto"]&&(window.location.href=data["goto"]),void 0!=data["function"]&&eval(data["function"]),void 0!=data.openDialog&&bootbox.dialog({message:data.content,title:data.message,icon:void 0!=data.icon?data.icon:"",width:void 0!=data.width?data.width:"",owner:void 0!=data.owner?data.owner:"",id:void 0!=data.windowId?data.windowId:"",animate:!1,closeButton:void 0!=data.closeButton?data.closeButton:!0}),void 0!=data.reload?void window.location.reload(!0):void 0},showLoader:function(){$(".cu-window-loader").hide(),$(".cu-window-blocker").hide();var t='<div class="cu-window-blocker"></div>';t+='<div class="cu-window-loader"><div class="loader-container text-center"><button class="close" type="button">&times;</button><span><span class="cu-icon cu-spinner" style="animation: cu-spin 1s infinite steps(10)"><svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14"><path d="M5.233 2.056C5.23 1.063 6.033.253 7.027.25c.994-.003 1.803.8 1.806 1.794.004.994-.8 1.802-1.794 1.806-1 .003-1.81-.8-1.81-1.794zm3.824 1.568c-.003-.993.8-1.802 1.792-1.806.99-.003 1.8.8 1.8 1.793 0 1-.8 1.81-1.8 1.81-1 .01-1.8-.8-1.81-1.79zm2.495 3.81c-.002-.497.398-.9.896-.903.497 0 .902.4.904.9 0 .5-.402.9-.897.91-.498.01-.902-.39-.903-.89zm-1.57 3.824c-.003-.496.4-.9.895-.903.498 0 .903.4.904.897.01.498-.4.9-.89.904-.49.002-.9-.4-.9-.898zM6.17 12.853c-.002-.496.4-.9.897-.902.497 0 .9.4.903.9.002.5-.4.9-.897.91-.498.01-.902-.4-.903-.89zm-3.824-1.57c-.002-.495.4-.9.897-.902.496 0 .9.4.902.9.002.5-.4.9-.896.91-.5 0-.91-.4-.91-.9zM1.87 3.65c-.003-.746.6-1.352 1.345-1.354.746-.002 1.352.6 1.354 1.345 0 .75-.6 1.36-1.35 1.36-.75.01-1.35-.6-1.36-1.34zM.637 7.472c-.002-.56.45-1.014 1.01-1.016.56-.002 1.013.45 1.015 1.01 0 .56-.45 1.014-1.008 1.016S.64 8.03.638 7.472z"/></svg></span> '+cuLanguage.inProgress+"</span></div></div>",$("body").append(t),$(".cu-window-blocker").fadeIn(0,function(){$(".cu-window-loader").fadeIn(1)})},closeLoader:function(t){$(".cu-window-loader").fadeOut(1,function(){$(".cu-window-blocker").fadeOut(0,function(){$(".cu-window-loader").remove(),$(".cu-window-blocker").remove()})})},getURI:function(t,e,o,n,i){return this.getControllerURI(t,e,o,n,i)},getControllerURI:function(module,controller,action,zone,params){var url=xoUrl;if(void 0!=cu_modules[module]&&void 0!=controller&&""!=controller)url+="backend"==zone?"/admin":"",url+=cu_modules[module]+"/"+controller+"/"+action+"/";else{if(url+="/modules/"+module,url+="backend"==zone?"/admin":"",""==controller||void 0==controller)return url;url+="/index.php/"+controller+"/"+action+"/"}if(void 0==params)return url;var query="";for(key in params)query+=(""==query?"?":"&")+key+"="+eval("params."+key);return url+query},url:function(t,e){var o=window.location.protocol+"//"+window.location.host;""!=window.location.port&&(o+=":"+window.location.port);var n=xoUrl.replace(o,"");return void 0==t?xoUrl:1==arguments.length||1!=e?xoUrl+t:n+t},createDataTable:function(t){if(!t.hasClass("dataTable")){var e=$(t).data("exclude");void 0!=e?e.toString().split(","):"";$(t).dataTable({bProcessing:!0,bServerSide:!0,sAjaxSource:$(t).data("source"),bPaginate:!0})}},runAction:function(e){var action=$(e).data("action");switch(action){case"load-remote-dialog":case"load-module-dialog":cuHandler.loadRemoteDialog(e);break;case"goto":var url=$(e).data("url");void 0==url&&(url=$(e).attr("href"));var retriever=$(e).data("retriever");void 0!=retriever&&(url+=eval(retriever+"(e)")),void 0!=url&&""!=url&&(window.location.href=url);break;case"ajax":var url=""!=$(e).attr("href")?$(e).attr("href"):$(e).data("url");cuHandler.requestAjax(e,{url:url})}},enableCommands:function(id_activator,type){var commands=$("*[data-activator='"+id_activator+"']"),total=$("#"+id_activator+" :"+type+"[data-switch]:checked").length;$(commands).each(function(index){var required=void 0!=$(this).data("oncount")?$(this).data("oncount"):" >= 1";eval("total "+required)?$(this).enable():$(this).disable()})},getIcon:function(t){var e="svg-"==t.slice(0,4),o=!1,n=!1;if(e||(e=".svg"==t.slice(-4),e&&(n=!0)),o=null!=t.match(/^(http:\/\/|ftp:\/\/|https:\/\/\|\/\/)/i),e&&!n){var i=t.split("-");if(i.length<3)return this.url("modules/rmcommon/icons/noicon.svg");var s=t.slice(5+i[1].length),r=this.url("modules/rmcommon/icons/"+s+".svg");if(void 0==iconsProviders)return r;if(iconsProviders.hasOwnProperty(i[1])){var a=iconsProviders[i[1]]+"/"+s+".svg";return a}return r}if(n&&!o)return this.url(t);var l=[".jpg",".gif",".png","jpeg"],c=t.slice(-4);return l.indexOf(c)||o?t:'<span class="'+t+'"></span>'},loadIcon:function(t,e,o){if(arguments.length<1)return!1;o=2==arguments.length?!0:1==arguments.length?!1:o;var n=this.getIcon(t),i=".svg"==n.slice(-4),s="<span"==n.slice(0,5);if(o)var r=e.find(".cu-icon");else var r=$("<span />",{"class":"cu-icon"});if((void 0==r||r.length<=0)&&(o=!1,r=$("<span />",{"class":"cu-icon"})),i)r.html("").load(n);else if(s)r.html("").append(n);else{var a=$("<img>").attr("src",n);r.html("").append(a)}return!o&&arguments.length>1&&$(e).append(r),r},modal:bootbox,notify:function(t){return new PNotify(t)}};Number.prototype.formatMoney=function(t,e,o){var n=this,t=isNaN(t=Math.abs(t))?2:t,e=void 0==e?".":e,o=void 0==o?",":o,i=0>n?"-":"",s=parseInt(n=Math.abs(+n||0).toFixed(t))+"",r=(r=s.length)>3?r%3:0;return i+(r?s.substr(0,r)+o:"")+s.substr(r).replace(/(\d{3})(?=\d)/g,"$1"+o)+(t?e+Math.abs(n-s).toFixed(t).slice(2):"")},jQuery.fn.enable=function(){this.each(function(){jQuery(this).removeAttr("disabled"),jQuery(this).removeClass("disabled")})},jQuery.fn.disable=function(){this.each(function(){jQuery(this).attr("disabled",!0),jQuery(this).addClass("disabled")})},$(document).ready(function(){var t="";if($("body").on("click","*[data-action]",function(t){return $(this).is(":disabled")||$(this).attr("disabled")||$(this).hasClass("disabled")?($(t).stopPropagation(),!1):(cuHandler.runAction($(this)),!1)}),$("body").on("submit",'form[data-type="ajax"]',function(){return cuHandler.submitAjaxForm($(this)),!1}),$("body").on("submit",'form[data-type="no-submit"]',function(){return!1}),$("body").on("click",".cu-window-loader .close",function(){cuHandler.closeLoader()}),$(".cu-data-table").each(function(){cuHandler.createDataTable($(this))}),$("*[data-rel='tooltip']").tooltip(),$("body").on("change",".activator-container :checkbox[data-switch], .activator-container :radio[data-switch]",function(){event.stopPropagation();var t=$(this).parents(".activator-container").attr("id");cuHandler.enableCommands(t,$(this).attr("type")),"radio"==$(this).attr("type")&&$(this).parents(".activator-container").find(".tr-checked").removeClass("tr-checked"),$(this).is(":checked")?$(this).parents("tr").addClass("tr-checked"):$(this).parents("tr").removeClass("tr-checked")}),$("body").on("click",".activator-container > tbody > tr",function(t){if("DIV"==t.target.tagName||"TD"==t.target.tagName){var e=$(this).find("input[data-switch]");"radio"==$(e).attr("type")&&$(this).parents(".activator-container").find(".tr-checked").removeClass("tr-checked"),e.is(":checked")?e.removeAttr("checked"):e.prop("checked","checked"),cuHandler.enableCommands($(this).parents(".activator-container").attr("id"),e.attr("type")),$(e).is(":checked")?$(this).addClass("tr-checked"):$(this).removeClass("tr-checked"),event.stopPropagation()}}),$("body").on("change",":checkbox[data-checkbox]",function(){var t=$(this).data("checkbox");if(void 0!=t){$(":checkbox[data-oncheck='"+t+"']").prop("checked",$(this).prop("checked")).change();var e=$(this).parents(".activator-container");e.length>0&&cuHandler.enableCommands(e.attr("id"),"checkbox")}}),$("body").on("change",":checkbox",function(){if(!this.hasAttribute("data-checkbox")&&this.hasAttribute("data-oncheck")){var t=$(":checkbox[data-oncheck='"+$(this).data("oncheck")+"']"),e=$(":checkbox[data-oncheck='"+$(this).data("oncheck")+"']:checked"),o=$(":checkbox[data-checkbox='"+$(this).data("oncheck")+"']");e.length<t.length?$(o).removeAttr("checked"):e.length==t.length&&$(o).prop("checked","checked")}}),1==$("*[data-news='load']").length||$("*[data-boxes='load']").length>0){var e=$("*[data-news='load']");if(e.length<=0&&(e=$("*[data-boxes='load']")),e.length<=0)return!1;var o=e.data("module"),n=$(e.data("target"));void 0!=n&&n.html('<div class="text-success"><span class="fa fa-spinner fa-pulse"></span> '+cuLanguage.downloadNews+"</div>");var s=$("*[data-boxes='load']"),r={module:o,CU_TOKEN:$("#cu-token").val()};$.get(xoUrl+"/modules/rmcommon/ajax/module-info.php",r,function(t){if("error"==t.type)return void n.html('<div class="text-danger"><span class="fa fa-exclamation-triangle"></span> '+cuLanguage.downloadNewsError+"</div>");if(void 0!=t.news&&void 0!=n){for(news=$("<ul>").addClass("cu-ajax-news list-unstyled"),i=0;i<t.news.length;i++){var e="<li><small>"+t.news[i].date+'</small><h5><a href="'+t.news[i].link+'" target="_blank">'+t.news[i].title+"</a></h5>";e+='<p class="help-block">'+t.news[i].content+"</p></li>",news.append(e)}n.html("").append(news),news.fadeIn("fast",function(){$('html.dashboard [data-container="dashboard"]').trigger("containerUpdated")})}if(void 0!=t.boxes&&void 0!=s)for(i=0;i<t.boxes.length;i++){if(void 0==t.boxes[i].size||t.boxes[i].size<=0)var o=1;else o=t.boxes[i].size;var r=$('<div data-dashboard="item">').addClass("size-"+o).css("display","none");if(r.append('<div class="cu-box"><div class="box-header"><span class="fa fa-caret-up box-handler"></span><h3 class="box-title">'+t.boxes[i].title+'</h3></div><div class="box-content">'+t.boxes[i].content+"</div></div>"),void 0!=t.boxes[i].container){var a=$(t.boxes[i].container);a.length>0&&$(a).each(function(){var e=r.clone();"top"==t.boxes[i].position?$(this).prepend(e):$(this).append(e),e.fadeIn("fast",function(){$('html.dashboard [data-container="dashboard"]').trigger("containerUpdated")})})}}},"json")}$("body").on("click",".ed-container .full-screen",function(){$(this).parents(".ed-container").addClass("full-screen-edit"),$(this).removeClass("full-screen").addClass("normal-screen"),t=$(this).parents(".ed-container").find(".txtarea-container").attr("style"),$(this).parents(".ed-container").find(".txtarea-container").attr("style",""),$("body").css("overflow","hidden")}),$("body").on("click",".ed-container .normal-screen",function(){$(this).parents(".ed-container").removeClass("full-screen-edit"),$(this).addClass("full-screen").removeClass("normal-screen"),$(this).parents(".ed-container").find(".txtarea-container").attr("style",t),t="",$("body").css("overflow","visible")}),$(".cu-notifications form .notification-item").change(function(){var t=$(this).data("event"),e=$(this).is(":checked")?1:0,o=$(this),n={event:t,status:e,page:"cu-notification-subscribe"};$.post(xoUrl+"/notifications.php",n,function(t){return"error"==t.type?void alert(t.message):void $(o).parent().parent().animate({backgroundColor:"#ffff99"},500,"linear",function(){$(this).animate({backgroundColor:"transparent"},500,"linear")})},"json")}),$(" #cu-notifications .cancel-subscription").click(function(){var t=$(this).data("info"),e=0,o=$(this),n={event:t,status:e,page:"cu-notification-subscribe"};$.post(xoUrl+"/notifications.php",n,function(t){return"error"==t.type?void alert(t.message):void $(o).parent().parent().slideUp(250,function(){$(this).remove()})},"json")})});
+!function(t,o){"use strict";"function"==typeof define&&define.amd?define(["jquery"],o):"object"==typeof exports?module.exports=o(require("jquery")):t.bootbox=o(t.jQuery)}(this,function t(o,e){"use strict";function a(t){var o=C[f.locale];return o?o[t]:C.en[t]}function n(t,e,a){t.stopPropagation(),t.preventDefault();var n=o.isFunction(a)&&a.call(e,t)===!1;n||e.modal("hide")}function r(t){var o,e=0;for(o in t)e++;return e}function i(t,e){var a=0;o.each(t,function(t,o){e(t,o,a++)})}function l(t){var e,a;if("object"!=typeof t)throw new Error("Please supply an object of options");if(!t.message)throw new Error("Please specify a message");return t=o.extend({},f,t),t.buttons||(t.buttons={}),e=t.buttons,a=r(e),i(e,function(t,n,r){if(o.isFunction(n)&&(n=e[t]={callback:n}),"object"!==o.type(n))throw new Error("button with key "+t+" must be an object");n.label||(n.label=t),n.className||(2>=a&&r===a-1?n.className="btn-primary":n.className="btn-default")}),t}function c(t,o){var e=t.length,a={};if(1>e||e>2)throw new Error("Invalid argument length");return 2===e||"string"==typeof t[0]?(a[o[0]]=t[0],a[o[1]]=t[1]):a=t[0],a}function s(t,e,a){return o.extend(!0,{},t,c(e,a))}function u(t,o,e,a){var n={className:"bootbox-"+t,buttons:p.apply(null,o)};return b(s(n,a,e),o)}function p(){for(var t={},o=0,e=arguments.length;e>o;o++){var n=arguments[o],r=n.toLowerCase(),i=n.toUpperCase();t[r]={label:a(i)}}return t}function b(t,o){var a={};return i(o,function(t,o){a[o]=!0}),i(t.buttons,function(t){if(a[t]===e)throw new Error("button key "+t+" is not allowed (options are "+o.join("\n")+")")}),t}var d={dialog:"<div class='bootbox modal' tabindex='-1' role='dialog'><div class='modal-dialog'><div class='modal-content'><div class='modal-body'><div class='bootbox-body'></div></div></div></div></div>",header:"<div class='modal-header'><h4 class='modal-title'></h4></div>",footer:"<div class='modal-footer'></div>",closeButton:"<button type='button' class='bootbox-close-button close' data-dismiss='modal' aria-hidden='true'>&times;</button>",form:"<form class='bootbox-form'></form>",inputs:{text:"<input class='bootbox-input bootbox-input-text form-control' autocomplete=off type=text />",textarea:"<textarea class='bootbox-input bootbox-input-textarea form-control'></textarea>",email:"<input class='bootbox-input bootbox-input-email form-control' autocomplete='off' type='email' />",select:"<select class='bootbox-input bootbox-input-select form-control'></select>",checkbox:"<div class='checkbox'><label><input class='bootbox-input bootbox-input-checkbox' type='checkbox' /></label></div>",date:"<input class='bootbox-input bootbox-input-date form-control' autocomplete=off type='date' />",time:"<input class='bootbox-input bootbox-input-time form-control' autocomplete=off type='time' />",number:"<input class='bootbox-input bootbox-input-number form-control' autocomplete=off type='number' />",password:"<input class='bootbox-input bootbox-input-password form-control' autocomplete='off' type='password' />"}},f={locale:"en",backdrop:"static",animate:!0,className:null,closeButton:!0,show:!0,container:"body",id:""},m={};m.alert=function(){var t;if(t=u("alert",["ok"],["message","callback"],arguments),t.callback&&!o.isFunction(t.callback))throw new Error("alert requires callback property to be a function when provided");return t.buttons.ok.callback=t.onEscape=function(){return o.isFunction(t.callback)?t.callback.call(this):!0},m.dialog(t)},m.confirm=function(){var t;if(t=u("confirm",["cancel","confirm"],["message","callback"],arguments),t.buttons.cancel.callback=t.onEscape=function(){return t.callback.call(this,!1)},t.buttons.confirm.callback=function(){return t.callback.call(this,!0)},!o.isFunction(t.callback))throw new Error("confirm requires a callback");return m.dialog(t)},m.prompt=function(){var t,a,n,r,l,c,u;if(r=o(d.form),a={className:"bootbox-prompt",buttons:p("cancel","confirm"),value:"",inputType:"text"},t=b(s(a,arguments,["title","callback"]),["cancel","confirm"]),c=t.show===e?!0:t.show,t.message=r,t.buttons.cancel.callback=t.onEscape=function(){return t.callback.call(this,null)},t.buttons.confirm.callback=function(){var e;switch(t.inputType){case"text":case"textarea":case"email":case"select":case"date":case"time":case"number":case"password":e=l.val();break;case"checkbox":var a=l.find("input:checked");e=[],i(a,function(t,a){e.push(o(a).val())})}return t.callback.call(this,e)},t.show=!1,!t.title)throw new Error("prompt requires a title");if(!o.isFunction(t.callback))throw new Error("prompt requires a callback");if(!d.inputs[t.inputType])throw new Error("invalid prompt type");switch(l=o(d.inputs[t.inputType]),t.inputType){case"text":case"textarea":case"email":case"date":case"time":case"number":case"password":l.val(t.value);break;case"select":var f={};if(u=t.inputOptions||[],!o.isArray(u))throw new Error("Please pass an array of input options");if(!u.length)throw new Error("prompt with select requires options");i(u,function(t,a){var n=l;if(a.value===e||a.text===e)throw new Error("given options in wrong format");a.group&&(f[a.group]||(f[a.group]=o("<optgroup/>").attr("label",a.group)),n=f[a.group]),n.append("<option value='"+a.value+"'>"+a.text+"</option>")}),i(f,function(t,o){l.append(o)}),l.val(t.value);break;case"checkbox":var C=o.isArray(t.value)?t.value:[t.value];if(u=t.inputOptions||[],!u.length)throw new Error("prompt with checkbox requires options");if(!u[0].value||!u[0].text)throw new Error("given options in wrong format");l=o("<div/>"),i(u,function(e,a){var n=o(d.inputs[t.inputType]);n.find("input").attr("value",a.value),n.find("label").append(a.text),i(C,function(t,o){o===a.value&&n.find("input").prop("checked",!0)}),l.append(n)})}return t.placeholder&&l.attr("placeholder",t.placeholder),t.pattern&&l.attr("pattern",t.pattern),t.maxlength&&l.attr("maxlength",t.maxlength),r.append(l),r.on("submit",function(t){t.preventDefault(),t.stopPropagation(),n.find(".btn-primary").click()}),n=m.dialog(t),n.off("shown.bs.modal"),n.on("shown.bs.modal",function(){l.focus()}),c===!0&&n.modal("show"),n},m.dialog=function(t){t=l(t);var a=o(d.dialog),r=a.find(".modal-dialog"),c=a.find(".modal-body"),s=t.buttons,u="",p={onEscape:t.onEscape};if(o.fn.modal===e)throw new Error("$.fn.modal is not defined; please double check you have included the Bootstrap JavaScript library. See http://getbootstrap.com/javascript/ for more details.");if(i(s,function(t,o){u+="<button data-bb-handler='"+t+"' type='button' class='btn "+o.className+"'>"+o.label+"</button>",p[t]=o.callback}),c.find(".bootbox-body").html(t.message),t.id&&a.attr("id",t.id),t.animate===!0&&a.addClass("fade"),t.owner&&a.attr("data-owner",t.owner),t.className&&a.addClass(t.className),t.width&&(t.size=t.width),t.size&&r.addClass(t.size),t.color!=e&&""!=t.color&&r.addClass("modal-"+t.color),t.title&&c.before(d.header),t.closeButton){var b=o(d.closeButton);t.title?a.find(".modal-header").prepend(b):b.css("margin-top","-10px").prependTo(c)}return t.title&&a.find(".modal-title").html(t.icon?'<i class="'+t.icon+'"></i> '+t.title:t.title),u.length&&(c.after(d.footer),a.find(".modal-footer").html(u)),a.on("hidden.bs.modal",function(t){t.target===this&&a.remove()}),a.on("shown.bs.modal",function(){a.find(".btn-primary:first").focus()}),"static"!==t.backdrop&&a.on("click.dismiss.bs.modal",function(t){a.children(".modal-backdrop").length&&(t.currentTarget=a.children(".modal-backdrop").get(0)),t.target===t.currentTarget&&a.trigger("escape.close.bb")}),""!=t.id&&a.attr("id",t.id),a.on("escape.close.bb",function(t){p.onEscape&&n(t,a,p.onEscape)}),a.on("click",".modal-footer button",function(t){var e=o(this).data("bb-handler");n(t,a,p[e])}),a.on("click",".bootbox-close-button",function(t){n(t,a,p.onEscape)}),a.on("keyup",function(t){27===t.which&&a.trigger("escape.close.bb")}),o(t.container).append(a),a.modal({backdrop:t.backdrop?"static":!1,keyboard:!1,show:!1}),t.show&&a.modal("show"),a},m.setDefaults=function(){var t={};2===arguments.length?t[arguments[0]]=arguments[1]:t=arguments[0],o.extend(f,t)},m.hideAll=function(){return o(".bootbox").modal("hide"),m};var C={bg_BG:{OK:"Ок",CANCEL:"Отказ",CONFIRM:"Потвърждавам"},br:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Sim"},cs:{OK:"OK",CANCEL:"Zrušit",CONFIRM:"Potvrdit"},da:{OK:"OK",CANCEL:"Annuller",CONFIRM:"Accepter"},de:{OK:"OK",CANCEL:"Abbrechen",CONFIRM:"Akzeptieren"},el:{OK:"Εντάξει",CANCEL:"Ακύρωση",CONFIRM:"Επιβεβαίωση"},en:{OK:"OK",CANCEL:"Cancel",CONFIRM:"OK"},es:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Aceptar"},et:{OK:"OK",CANCEL:"Katkesta",CONFIRM:"OK"},fa:{OK:"قبول",CANCEL:"لغو",CONFIRM:"تایید"},fi:{OK:"OK",CANCEL:"Peruuta",CONFIRM:"OK"},fr:{OK:"OK",CANCEL:"Annuler",CONFIRM:"D'accord"},he:{OK:"אישור",CANCEL:"ביטול",CONFIRM:"אישור"},hu:{OK:"OK",CANCEL:"Mégsem",CONFIRM:"Megerősít"},hr:{OK:"OK",CANCEL:"Odustani",CONFIRM:"Potvrdi"},id:{OK:"OK",CANCEL:"Batal",CONFIRM:"OK"},it:{OK:"OK",CANCEL:"Annulla",CONFIRM:"Conferma"},ja:{OK:"OK",CANCEL:"キャンセル",CONFIRM:"確認"},lt:{OK:"Gerai",CANCEL:"Atšaukti",CONFIRM:"Patvirtinti"},lv:{OK:"Labi",CANCEL:"Atcelt",CONFIRM:"Apstiprināt"},nl:{OK:"OK",CANCEL:"Annuleren",CONFIRM:"Accepteren"},no:{OK:"OK",CANCEL:"Avbryt",CONFIRM:"OK"},pl:{OK:"OK",CANCEL:"Anuluj",CONFIRM:"Potwierdź"},pt:{OK:"OK",CANCEL:"Cancelar",CONFIRM:"Confirmar"},ru:{OK:"OK",CANCEL:"Отмена",CONFIRM:"Применить"},sq:{OK:"OK",CANCEL:"Anulo",CONFIRM:"Prano"},sv:{OK:"OK",CANCEL:"Avbryt",CONFIRM:"OK"},th:{OK:"ตกลง",CANCEL:"ยกเลิก",CONFIRM:"ยืนยัน"},tr:{OK:"Tamam",CANCEL:"İptal",CONFIRM:"Onayla"},zh_CN:{OK:"OK",CANCEL:"取消",CONFIRM:"确认"},zh_TW:{OK:"OK",CANCEL:"取消",CONFIRM:"確認"}};return m.addLocale=function(t,e){return o.each(["OK","CANCEL","CONFIRM"],function(t,o){if(!e[o])throw new Error("Please supply a translation for '"+o+"'")}),C[t]={OK:e.OK,CANCEL:e.CANCEL,CONFIRM:e.CONFIRM},m},m.removeLocale=function(t){return delete C[t],m},m.setLocale=function(t){return m.setDefaults("locale",t)},m.init=function(e){return t(e||o)},m});
+!function(t){t.fn.cuSpinner=function(e){var o=t.extend({icon:"svg-rmcommon-spinner-06",type:"spin",hide:"cu-icon",steps:12,speed:3},e);o.type="spin"!=o.type&&"pulse"!=o.type?"spin":o.type;"pulse"==o.type?"animation: cu-spin 1s infinite steps("+o.steps+")":"animation: cu-spin "+2/6*o.speed+"s infinite linear;";return this.each(function(){var e=t(this),n=cuHandler.getIcon(o.icon);n||(n=cuHandler.getIcon("svg-rmcommon-spinner-06"));var a=t(this).find(".cu-spinner");if(a.length>0)return t(a).remove(),t(this).find(".cu-spinner-hide").removeClass("cu-spinner-hide").show(),!0;var r=t("<span />",{"class":"cu-icon cu-spinner"});return r.load(n,function(){var n=t(e).find("."+o.hide);n.length>0?(n.addClass("cu-spinner-hide").hide(),t(n[0]).after(r)):t(e).prepend(r)}),!0})}}(jQuery);
+!function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery")):"function"==typeof define&&define.amd?define("pnotify",["jquery"],t):t(jQuery)}(function(t){var o,e,i={dir1:"down",dir2:"left",push:"bottom",spacing1:25,spacing2:25,context:t("body")},n=t(window),s=function(){e=t("body"),PNotify.prototype.options.stack.context=e,n=t(window),n.bind("resize",function(){o&&clearTimeout(o),o=setTimeout(function(){PNotify.positionAll(!0)},10)})};return PNotify=function(t){this.parseOptions(t),this.init()},t.extend(PNotify.prototype,{version:"2.1.0",options:{title:!1,title_escape:!1,text:!1,text_escape:!1,styling:"brighttheme",addclass:"",cornerclass:"",auto_display:!0,width:"300px",min_height:"16px",type:"notice",icon:!0,opacity:1,animation:"fade",animate_speed:"slow",position_animate_speed:500,shadow:!0,hide:!0,delay:8e3,mouse_reset:!0,remove:!0,insert_brs:!0,destroy:!0,stack:i},modules:{},runModules:function(t,o){var e;for(var i in this.modules)e="object"==typeof o&&i in o?o[i]:o,"function"==typeof this.modules[i][t]&&this.modules[i][t](this,"object"==typeof this.options[i]?this.options[i]:{},e)},state:"initializing",timer:null,styles:null,elem:null,container:null,title_container:null,text_container:null,animating:!1,timerHide:!1,init:function(){var o=this;if(this.modules={},t.extend(!0,this.modules,PNotify.prototype.modules),"object"==typeof this.options.styling?this.styles=this.options.styling:this.styles=PNotify.styling[this.options.styling],this.elem=t("<div />",{"class":"ui-pnotify "+this.options.addclass,css:{display:"none"},"aria-live":"assertive",mouseenter:function(t){if(o.options.mouse_reset&&"out"===o.animating){if(!o.timerHide)return;o.cancelRemove()}o.options.hide&&o.options.mouse_reset&&o.cancelRemove()},mouseleave:function(t){o.options.hide&&o.options.mouse_reset&&"out"!==o.animating&&o.queueRemove(),PNotify.positionAll()}}),this.container=t("<div />",{"class":"alert "+this.options.type,role:"alert"}).appendTo(this.elem),""!==this.options.cornerclass&&this.container.removeClass("ui-corner-all").addClass(this.options.cornerclass),this.options.shadow&&this.container.addClass("ui-pnotify-shadow"),this.options.icon!==!1)if("string"==typeof this.options.icon){var e=t("<div />",{"class":"ui-pnotify-icon"});cuHandler.loadIcon(this.options.icon,e),e.prependTo(this.container)}else t("<div />",{"class":"ui-pnotify-icon"}).append(t("<span />",{"class":this.options.icon===!0?"error"===this.options.type?this.styles.error_icon:"info"===this.options.type?this.styles.info_icon:"success"===this.options.type?this.styles.success_icon:this.styles.notice_icon:this.options.icon})).prependTo(this.container);return this.title_container=t("<h4 />",{"class":"ui-pnotify-title"}).appendTo(this.container),this.options.title===!1?this.title_container.hide():this.options.title_escape?this.title_container.text(this.options.title):this.title_container.html(this.options.title),this.text_container=t("<div />",{"class":"ui-pnotify-text"}).appendTo(this.container),this.options.text===!1?this.text_container.hide():this.options.text_escape?this.text_container.text(this.options.text):this.text_container.html(this.options.insert_brs?String(this.options.text).replace(/\n/g,"<br />"):this.options.text),"string"==typeof this.options.min_height&&this.container.css("min-height",this.options.min_height),"top"===this.options.stack.push?PNotify.notices=t.merge([this],PNotify.notices):PNotify.notices=t.merge(PNotify.notices,[this]),"top"===this.options.stack.push&&this.queuePosition(!1,1),this.options.stack.animation=!1,this.runModules("init"),this.options.auto_display&&this.open(),this},update:function(o){var e=this.options;if(this.parseOptions(e,o),this.options.cornerclass!==e.cornerclass&&this.container.removeClass("ui-corner-all "+e.cornerclass).addClass(this.options.cornerclass),this.options.shadow!==e.shadow&&(this.options.shadow?this.container.addClass("ui-pnotify-shadow"):this.container.removeClass("ui-pnotify-shadow")),this.options.addclass===!1?this.elem.removeClass(e.addclass):this.options.addclass!==e.addclass&&this.elem.removeClass(e.addclass).addClass(this.options.addclass),this.options.title===!1?this.title_container.slideUp("fast"):this.options.title!==e.title&&(this.options.title_escape?this.title_container.text(this.options.title):this.title_container.html(this.options.title),e.title===!1&&this.title_container.slideDown(200)),this.options.text===!1?this.text_container.slideUp("fast"):this.options.text!==e.text&&(this.options.text_escape?this.text_container.text(this.options.text):this.text_container.html(this.options.insert_brs?String(this.options.text).replace(/\n/g,"<br />"):this.options.text),e.text===!1&&this.text_container.slideDown(200)),this.options.type!==e.type&&this.container.removeClass(this.styles.error+" "+this.styles.notice+" "+this.styles.success+" "+this.styles.info).addClass(this.options.type),(this.options.icon!==e.icon||this.options.icon===!0&&this.options.type!==e.type)&&(this.container.find("div.ui-pnotify-icon").remove(),this.options.icon!==!1))if("string"==typeof this.options.icon){var i=t("<div />",{"class":"ui-pnotify-icon"});cuHandler.loadIcon(this.options.icon,i),i.prependTo(this.container)}else t("<div />",{"class":"ui-pnotify-icon"}).append(t("<span />",{"class":this.options.icon===!0?"error"===this.options.type?this.styles.error_icon:"info"===this.options.type?this.styles.info_icon:"success"===this.options.type?this.styles.success_icon:this.styles.notice_icon:this.options.icon})).prependTo(this.container);return this.options.min_height!==e.min_height&&this.container.animate({minHeight:this.options.min_height}),this.options.opacity!==e.opacity&&this.elem.fadeTo(this.options.animate_speed,this.options.opacity),this.options.hide?e.hide||this.queueRemove():this.cancelRemove(),this.queuePosition(!0),this.runModules("update",e),this},open:function(){this.state="opening",this.runModules("beforeOpen");var t=this;return this.elem.parent().length||this.elem.appendTo(this.options.stack.context?this.options.stack.context:e),"top"!==this.options.stack.push&&this.position(!0),"fade"===this.options.animation||"fade"===this.options.animation.effect_in?this.elem.show().fadeTo(0,0).hide():1!==this.options.opacity&&this.elem.show().fadeTo(0,this.options.opacity).hide(),this.animateIn(function(){t.queuePosition(!0),t.options.hide&&t.queueRemove(),t.state="open",t.runModules("afterOpen")}),this},remove:function(o){this.state="closing",this.timerHide=!!o,this.runModules("beforeClose");var e=this;return this.timer&&(window.clearTimeout(this.timer),this.timer=null),this.animateOut(function(){if(e.state="closed",e.runModules("afterClose"),e.queuePosition(!0),e.options.remove&&e.elem.detach(),e.runModules("beforeDestroy"),e.options.destroy&&null!==PNotify.notices){var o=t.inArray(e,PNotify.notices);-1!==o&&PNotify.notices.splice(o,1)}e.runModules("afterDestroy")}),this},get:function(){return this.elem},parseOptions:function(o,e){this.options=t.extend(!0,{},PNotify.prototype.options),this.options.stack=PNotify.prototype.options.stack;for(var i,n=[o,e],s=0;s<n.length&&(i=n[s],"undefined"!=typeof i);s++)if("object"!=typeof i)this.options.text=i;else for(var c in i)this.modules[c]?t.extend(!0,this.options[c],i[c]):this.options[c]=i[c]},animateIn:function(t){this.animating="in";var o;o="undefined"!=typeof this.options.animation.effect_in?this.options.animation.effect_in:this.options.animation,"none"===o?(this.elem.show(),t()):"show"===o?this.elem.show(this.options.animate_speed,t):"fade"===o?this.elem.show().fadeTo(this.options.animate_speed,this.options.opacity,t):"slide"===o?this.elem.slideDown(this.options.animate_speed,t):"function"==typeof o?o("in",t,this.elem):this.elem.show(o,"object"==typeof this.options.animation.options_in?this.options.animation.options_in:{},this.options.animate_speed,t),this.elem.parent().hasClass("ui-effects-wrapper")&&this.elem.parent().css({position:"fixed",overflow:"visible"}),"slide"!==o&&this.elem.css("overflow","visible"),this.container.css("overflow","hidden")},animateOut:function(t){this.animating="out";var o;o="undefined"!=typeof this.options.animation.effect_out?this.options.animation.effect_out:this.options.animation,"none"===o?(this.elem.hide(),t()):"show"===o?this.elem.hide(this.options.animate_speed,t):"fade"===o?this.elem.fadeOut(this.options.animate_speed,t):"slide"===o?this.elem.slideUp(this.options.animate_speed,t):"function"==typeof o?o("out",t,this.elem):this.elem.hide(o,"object"==typeof this.options.animation.options_out?this.options.animation.options_out:{},this.options.animate_speed,t),this.elem.parent().hasClass("ui-effects-wrapper")&&this.elem.parent().css({position:"fixed",overflow:"visible"}),"slide"!==o&&this.elem.css("overflow","visible"),this.container.css("overflow","hidden")},position:function(t){var o=this.options.stack,i=this.elem;if(i.parent().hasClass("ui-effects-wrapper")&&(i=this.elem.css({left:"0",top:"0",right:"0",bottom:"0"}).parent()),"undefined"==typeof o.context&&(o.context=e),o){"number"!=typeof o.nextpos1&&(o.nextpos1=o.firstpos1),"number"!=typeof o.nextpos2&&(o.nextpos2=o.firstpos2),"number"!=typeof o.addpos2&&(o.addpos2=0);var s="none"===i.css("display");if(!s||t){var c,a,r,l={};switch(o.dir1){case"down":r="top";break;case"up":r="bottom";break;case"left":r="right";break;case"right":r="left"}c=parseInt(i.css(r).replace(/(?:\..*|[^0-9.])/g,"")),isNaN(c)&&(c=0),"undefined"!=typeof o.firstpos1||s||(o.firstpos1=c,o.nextpos1=o.firstpos1);var p;switch(o.dir2){case"down":p="top";break;case"up":p="bottom";break;case"left":p="right";break;case"right":p="left"}if(a=parseInt(i.css(p).replace(/(?:\..*|[^0-9.])/g,"")),isNaN(a)&&(a=0),"undefined"!=typeof o.firstpos2||s||(o.firstpos2=a,o.nextpos2=o.firstpos2),("down"===o.dir1&&o.nextpos1+i.height()>(o.context.is(e)?n.height():o.context.prop("scrollHeight"))||"up"===o.dir1&&o.nextpos1+i.height()>(o.context.is(e)?n.height():o.context.prop("scrollHeight"))||"left"===o.dir1&&o.nextpos1+i.width()>(o.context.is(e)?n.width():o.context.prop("scrollWidth"))||"right"===o.dir1&&o.nextpos1+i.width()>(o.context.is(e)?n.width():o.context.prop("scrollWidth")))&&(o.nextpos1=o.firstpos1,o.nextpos2+=o.addpos2+("undefined"==typeof o.spacing2?25:o.spacing2),o.addpos2=0),o.animation&&o.nextpos2<a)switch(o.dir2){case"down":l.top=o.nextpos2+"px";break;case"up":l.bottom=o.nextpos2+"px";break;case"left":l.right=o.nextpos2+"px";break;case"right":l.left=o.nextpos2+"px"}else"number"==typeof o.nextpos2&&i.css(p,o.nextpos2+"px");switch(o.dir2){case"down":case"up":i.outerHeight(!0)>o.addpos2&&(o.addpos2=i.height());break;case"left":case"right":i.outerWidth(!0)>o.addpos2&&(o.addpos2=i.width())}if("number"==typeof o.nextpos1)if(o.animation&&(c>o.nextpos1||l.top||l.bottom||l.right||l.left))switch(o.dir1){case"down":l.top=o.nextpos1+"px";break;case"up":l.bottom=o.nextpos1+"px";break;case"left":l.right=o.nextpos1+"px";break;case"right":l.left=o.nextpos1+"px"}else i.css(r,o.nextpos1+"px");switch((l.top||l.bottom||l.right||l.left)&&i.animate(l,{duration:this.options.position_animate_speed,queue:!1}),o.dir1){case"down":case"up":o.nextpos1+=i.height()+("undefined"==typeof o.spacing1?25:o.spacing1);break;case"left":case"right":o.nextpos1+=i.width()+("undefined"==typeof o.spacing1?25:o.spacing1)}}return this}},queuePosition:function(t,e){return o&&clearTimeout(o),e||(e=10),o=setTimeout(function(){PNotify.positionAll(t)},e),this},cancelRemove:function(){return this.timer&&window.clearTimeout(this.timer),"closing"===this.state&&(this.elem.stop(!0),this.state="open",this.animating="in",this.elem.css("height","auto").animate({width:this.options.width,opacity:this.options.opacity},"fast")),this},queueRemove:function(){var t=this;return this.cancelRemove(),this.timer=window.setTimeout(function(){t.remove(!0)},isNaN(this.options.delay)?0:this.options.delay),this}}),t.extend(PNotify,{notices:[],removeAll:function(){t.each(PNotify.notices,function(){this.remove&&this.remove(!1)})},positionAll:function(e){if(o&&clearTimeout(o),o=null,PNotify.notices&&PNotify.notices.length)t.each(PNotify.notices,function(){var t=this.options.stack;t&&(t.nextpos1=t.firstpos1,t.nextpos2=t.firstpos2,t.addpos2=0,t.animation=e)}),t.each(PNotify.notices,function(){this.position()});else{var i=PNotify.prototype.options.stack;i&&(delete i.nextpos1,delete i.nextpos2)}},styling:{brighttheme:{container:"brighttheme",notice:"brighttheme-notice",notice_icon:"brighttheme-icon-notice",info:"brighttheme-info",info_icon:"brighttheme-icon-info",success:"brighttheme-success",success_icon:"brighttheme-icon-success",error:"brighttheme-error",error_icon:"brighttheme-icon-error"},jqueryui:{container:"ui-widget ui-widget-content ui-corner-all",notice:"ui-state-highlight",notice_icon:"ui-icon ui-icon-info",info:"",info_icon:"ui-icon ui-icon-info",success:"ui-state-default",success_icon:"ui-icon ui-icon-circle-check",error:"ui-state-error",error_icon:"ui-icon ui-icon-alert"},bootstrap2:{container:"alert",notice:"",notice_icon:"icon-exclamation-sign",info:"alert-info",info_icon:"icon-info-sign",success:"alert-success",success_icon:"icon-ok-sign",error:"alert-error",error_icon:"icon-warning-sign"},bootstrap3:{container:"alert",notice:"alert-warning",notice_icon:"glyphicon glyphicon-exclamation-sign",info:"alert-info",info_icon:"glyphicon glyphicon-info-sign",success:"alert-success",success_icon:"glyphicon glyphicon-ok-sign",error:"alert-danger",error_icon:"glyphicon glyphicon-warning-sign"}}}),PNotify.styling.fontawesome=t.extend({},PNotify.styling.bootstrap3),t.extend(PNotify.styling.fontawesome,{notice_icon:"fa fa-exclamation-circle",info_icon:"fa fa-info",success_icon:"fa fa-check",error_icon:"fa fa-warning"}),document.body?s():t(s),PNotify}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.buttons",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,o){o.prototype.options.buttons={closer:!0,closer_hover:!0,sticker:!0,sticker_hover:!0,show_on_nonblock:!1,labels:{close:"Close",stick:"Stick"}},o.prototype.modules.buttons={myOptions:null,closer:null,sticker:null,init:function(o,e){var i=this;this.myOptions=e,o.elem.on({mouseenter:function(t){!i.myOptions.sticker||o.options.nonblock&&o.options.nonblock.nonblock&&!i.myOptions.show_on_nonblock||i.sticker.trigger("pnotify_icon").css("visibility","visible"),!i.myOptions.closer||o.options.nonblock&&o.options.nonblock.nonblock&&!i.myOptions.show_on_nonblock||i.closer.css("visibility","visible")},mouseleave:function(t){i.myOptions.sticker_hover&&i.sticker.css("visibility","hidden"),i.myOptions.closer_hover&&i.closer.css("visibility","hidden")}}),this.sticker=t("<div />",{"class":"ui-pnotify-sticker",css:{cursor:"pointer",visibility:e.sticker_hover?"hidden":"visible"},click:function(){o.options.hide=!o.options.hide,o.options.hide?o.queueRemove():o.cancelRemove(),t(this).trigger("pnotify_icon")}}).bind("pnotify_icon",function(){t(this).children().removeClass(o.styles.pin_up+" "+o.styles.pin_down).addClass(o.options.hide?o.styles.pin_up:o.styles.pin_down),t(this).children().load(cuHandler.url("/modules/rmcommon/icons/"+(o.options.hide?"pause.svg":"play.svg")))}).append(t("<span />",{"class":o.styles.pin_up,title:e.labels.stick}).load(cuHandler.url("/modules/rmcommon/icons/pause.svg"))).prependTo(o.container),(!e.sticker||o.options.nonblock&&o.options.nonblock.nonblock&&!e.show_on_nonblock)&&this.sticker.css("display","none"),this.closer=t("<div />",{"class":"ui-pnotify-closer",css:{cursor:"pointer",visibility:e.closer_hover?"hidden":"visible"},click:function(){o.remove(!1),i.sticker.css("visibility","hidden"),i.closer.css("visibility","hidden")}}).append(t("<span />",{"class":o.styles.closer,title:e.labels.close}).load(cuHandler.url("/modules/rmcommon/icons/close.svg"))).prependTo(o.container),(!e.closer||o.options.nonblock&&o.options.nonblock.nonblock&&!e.show_on_nonblock)&&this.closer.css("display","none")},update:function(t,o){this.myOptions=o,!o.closer||t.options.nonblock&&t.options.nonblock.nonblock&&!o.show_on_nonblock?this.closer.css("display","none"):o.closer&&this.closer.css("display","block"),!o.sticker||t.options.nonblock&&t.options.nonblock.nonblock&&!o.show_on_nonblock?this.sticker.css("display","none"):o.sticker&&this.sticker.css("display","block"),this.sticker.trigger("pnotify_icon"),o.sticker_hover?this.sticker.css("visibility","hidden"):t.options.nonblock&&t.options.nonblock.nonblock&&!o.show_on_nonblock||this.sticker.css("visibility","visible"),o.closer_hover?this.closer.css("visibility","hidden"):t.options.nonblock&&t.options.nonblock.nonblock&&!o.show_on_nonblock||this.closer.css("visibility","visible")}},t.extend(o.styling.brighttheme,{closer:"cu-icon",pin_up:"cu-icon",pin_down:"cu-icon"}),t.extend(o.styling.jqueryui,{closer:"ui-icon ui-icon-close",pin_up:"ui-icon ui-icon-pin-w",pin_down:"ui-icon ui-icon-pin-s"}),t.extend(o.styling.bootstrap2,{closer:"icon-remove",pin_up:"icon-pause",pin_down:"icon-play"}),t.extend(o.styling.bootstrap3,{closer:"glyphicon glyphicon-remove",pin_up:"glyphicon glyphicon-pause",pin_down:"glyphicon glyphicon-play"}),t.extend(o.styling.fontawesome,{closer:"fa fa-times",pin_up:"fa fa-pause",pin_down:"fa fa-play"})}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.callbacks",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,o){var e=o.prototype.init,i=o.prototype.open,n=o.prototype.remove;o.prototype.init=function(){this.options.before_init&&this.options.before_init(this.options),e.apply(this,arguments),this.options.after_init&&this.options.after_init(this)},o.prototype.open=function(){var t;this.options.before_open&&(t=this.options.before_open(this)),t!==!1&&(i.apply(this,arguments),this.options.after_open&&this.options.after_open(this))},o.prototype.remove=function(t){var o;this.options.before_close&&(o=this.options.before_close(this,t)),o!==!1&&(n.apply(this,arguments),this.options.after_close&&this.options.after_close(this,t))}}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.confirm",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,o){o.prototype.options.confirm={confirm:!1,prompt:!1,prompt_class:"",prompt_default:"",prompt_multi_line:!1,align:"right",buttons:[{text:"Ok",addClass:"",promptTrigger:!0,click:function(t,o){t.remove(),t.get().trigger("pnotify.confirm",[t,o])}},{text:"Cancel",addClass:"",click:function(t){t.remove(),t.get().trigger("pnotify.cancel",t)}}]},o.prototype.modules.confirm={container:null,prompt:null,init:function(o,e){this.container=t('<div style="margin-top:5px;clear:both;" />').css("text-align",e.align).appendTo(o.container),e.confirm||e.prompt?this.makeDialog(o,e):this.container.hide()},update:function(t,o){o.confirm?(this.makeDialog(t,o),this.container.show()):this.container.hide().empty()},afterOpen:function(t,o){o.prompt&&this.prompt.focus()},makeDialog:function(o,e){var i,n,s=!1,c=this;this.container.empty(),e.prompt&&(this.prompt=t("<"+(e.prompt_multi_line?'textarea rows="5"':'input type="text"')+' style="margin-bottom:5px;clear:both;" />').addClass(o.styles.input+" "+e.prompt_class).val(e.prompt_default).appendTo(this.container));for(var a in e.buttons)i=e.buttons[a],s?this.container.append(" "):s=!0,n=t('<button type="button" />').addClass(o.styles.btn+" "+i.addClass).text(i.text).appendTo(this.container).on("click",function(t){return function(){"function"==typeof t.click&&t.click(o,e.prompt?c.prompt.val():null)}}(i)),e.prompt&&!e.prompt_multi_line&&i.promptTrigger&&this.prompt.keypress(function(t){return function(o){13==o.keyCode&&t.click()}}(n)),o.styles.text&&n.wrapInner('<span class="'+o.styles.text+'"></span>'),o.styles.btnhover&&n.hover(function(t){return function(){t.addClass(o.styles.btnhover)}}(n),function(t){return function(){t.removeClass(o.styles.btnhover)}}(n)),o.styles.btnactive&&n.on("mousedown",function(t){return function(){t.addClass(o.styles.btnactive)}}(n)).on("mouseup",function(t){return function(){t.removeClass(o.styles.btnactive)}}(n)),o.styles.btnfocus&&n.on("focus",function(t){return function(){t.addClass(o.styles.btnfocus)}}(n)).on("blur",function(t){return function(){t.removeClass(o.styles.btnfocus)}}(n))}},t.extend(o.styling.jqueryui,{btn:"ui-button ui-widget ui-state-default ui-corner-all ui-button-text-only",btnhover:"ui-state-hover",btnactive:"ui-state-active",btnfocus:"ui-state-focus",input:"",text:"ui-button-text"}),t.extend(o.styling.bootstrap2,{btn:"btn",input:""}),t.extend(o.styling.bootstrap3,{btn:"btn btn-default",input:"form-control"}),t.extend(o.styling.fontawesome,{btn:"btn btn-default",input:"form-control"})}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.desktop",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,o){var e,i=function(t,o){return(i="Notification"in window?function(t,o){return new Notification(t,o)}:"mozNotification"in navigator?function(t,o){return navigator.mozNotification.createNotification(t,o.body,o.icon).show()}:"webkitNotifications"in window?function(t,o){return window.webkitNotifications.createNotification(o.icon,t,o.body)}:function(t,o){return null})(t,o)};o.prototype.options.desktop={desktop:!1,fallback:!0,icon:null,tag:null},o.prototype.modules.desktop={tag:null,icon:null,genNotice:function(t,o){null===o.icon?this.icon="http://sciactive.com/pnotify/includes/desktop/"+t.options.type+".png":o.icon===!1?this.icon=null:this.icon=o.icon,(null===this.tag||null!==o.tag)&&(this.tag=null===o.tag?"PNotify-"+Math.round(1e6*Math.random()):o.tag),t.desktop=i(t.options.title,{icon:this.icon,body:t.options.text,tag:this.tag}),!("close"in t.desktop)&&"cancel"in t.desktop&&(t.desktop.close=function(){t.desktop.cancel()}),t.desktop.onclick=function(){t.elem.trigger("click")},t.desktop.onclose=function(){"closing"!==t.state&&"closed"!==t.state&&t.remove()}},init:function(t,i){return i.desktop?(e=o.desktop.checkPermission(),0!==e?void(i.fallback||(t.options.auto_display=!1)):void this.genNotice(t,i)):void 0},update:function(t,o,i){0!==e&&o.fallback||!o.desktop||this.genNotice(t,o)},beforeOpen:function(t,o){0!==e&&o.fallback||!o.desktop||t.elem.css({left:"-10000px",display:"none"})},afterOpen:function(t,o){0!==e&&o.fallback||!o.desktop||(t.elem.css({left:"-10000px",display:"none"}),"show"in t.desktop&&t.desktop.show())},beforeClose:function(t,o){0!==e&&o.fallback||!o.desktop||t.elem.css({left:"-10000px",display:"none"})},afterClose:function(t,o){0!==e&&o.fallback||!o.desktop||(t.elem.css({left:"-10000px",display:"none"}),"close"in t.desktop&&t.desktop.close())}},o.desktop={permission:function(){"undefined"!=typeof Notification&&"requestPermission"in Notification?Notification.requestPermission():"webkitNotifications"in window&&window.webkitNotifications.requestPermission()},checkPermission:function(){return"undefined"!=typeof Notification&&"permission"in Notification?"granted"===Notification.permission?0:1:"webkitNotifications"in window&&0==window.webkitNotifications.checkPermission()?0:1}},e=o.desktop.checkPermission()}),function(t){"object"==typeof exports&&"undefined"!=typeof module?module.exports=t(require("jquery"),require("pnotify")):"function"==typeof define&&define.amd?define("pnotify.nonblock",["jquery","pnotify"],t):t(jQuery,PNotify)}(function(t,o){var e,i=/^on/,n=/^(dbl)?click$|^mouse(move|down|up|over|out|enter|leave)$|^contextmenu$/,s=/^(focus|blur|select|change|reset)$|^key(press|down|up)$/,c=/^(scroll|resize|(un)?load|abort|error)$/,a=function(o,e){var a;if(o=o.toLowerCase(),document.createEvent&&this.dispatchEvent){if(o=o.replace(i,""),o.match(n)?(t(this).offset(),a=document.createEvent("MouseEvents"),a.initMouseEvent(o,e.bubbles,e.cancelable,e.view,e.detail,e.screenX,e.screenY,e.clientX,e.clientY,e.ctrlKey,e.altKey,e.shiftKey,e.metaKey,e.button,e.relatedTarget)):o.match(s)?(a=document.createEvent("UIEvents"),a.initUIEvent(o,e.bubbles,e.cancelable,e.view,e.detail)):o.match(c)&&(a=document.createEvent("HTMLEvents"),a.initEvent(o,e.bubbles,e.cancelable)),!a)return;this.dispatchEvent(a)}else o.match(i)||(o="on"+o),a=document.createEventObject(e),this.fireEvent(o,a)},r=function(o,i,n){o.elem.css("display","none");var s=document.elementFromPoint(i.clientX,i.clientY);o.elem.css("display","block");var c=t(s),r=c.css("cursor");"auto"===r&&"A"===s.tagName&&(r="pointer"),o.elem.css("cursor","auto"!==r?r:"default"),e&&e.get(0)==s||(e&&(a.call(e.get(0),"mouseleave",i.originalEvent),a.call(e.get(0),"mouseout",i.originalEvent)),a.call(s,"mouseenter",i.originalEvent),a.call(s,"mouseover",i.originalEvent)),a.call(s,n,i.originalEvent),e=c};o.prototype.options.nonblock={nonblock:!1,nonblock_opacity:.2},o.prototype.modules.nonblock={myOptions:null,init:function(t,o){var i=this;this.myOptions=o,t.elem.on({mouseenter:function(o){i.myOptions.nonblock&&o.stopPropagation(),i.myOptions.nonblock&&t.elem.stop().animate({opacity:i.myOptions.nonblock_opacity},"fast")},mouseleave:function(o){i.myOptions.nonblock&&o.stopPropagation(),e=null,t.elem.css("cursor","auto"),i.myOptions.nonblock&&"out"!==t.animating&&t.elem.stop().animate({opacity:t.options.opacity},"fast")},mouseover:function(t){i.myOptions.nonblock&&t.stopPropagation()},mouseout:function(t){i.myOptions.nonblock&&t.stopPropagation()},mousemove:function(o){i.myOptions.nonblock&&(o.stopPropagation(),r(t,o,"onmousemove"))},mousedown:function(o){i.myOptions.nonblock&&(o.stopPropagation(),o.preventDefault(),r(t,o,"onmousedown"))},mouseup:function(o){i.myOptions.nonblock&&(o.stopPropagation(),o.preventDefault(),r(t,o,"onmouseup"))},click:function(o){i.myOptions.nonblock&&(o.stopPropagation(),r(t,o,"onclick"))},dblclick:function(o){i.myOptions.nonblock&&(o.stopPropagation(),r(t,o,"ondblclick"))}})},update:function(t,o){this.myOptions=o}}});
+var cuHandler={ismobile:/android|webos|iphone|ipad|ipod|blackberry|iemobile|opera mini/i.test(navigator.userAgent.toLowerCase()),currentResponse:!1,loadRemoteDialog:function(launcher){var url=void 0!=$(launcher).attr("href")&&"#"!=$(launcher).attr("href")?$(launcher).attr("href"):$(launcher).data("url"),handler=$(launcher).data("handler"),window_id=$(launcher).data("window-id"),params={CUTOKEN_REQUEST:$("#cu-token").val()};return 0==params?!1:(cuHandler.showLoader(),$.get(url,params,function(response){return cuHandler.retrieveAjax(response)?(void 0==handler&&""!=handler?(cuHandler.closeLoader(),bootbox.dialog({message:response.content,title:response.message,icon:void 0!=response.icon?response.icon:"",width:void 0!=response.width?response.width:"",id:void 0!=response.id?response.id:window_id,animate:!1,color:void 0!=response.color?response.color:"",closeButton:void 0!=response.closeButton?response.closeButton:!0})):(cuHandler.closeLoader(),eval(handler+"(response, launcher);")),$(".cu-data-table").each(function(){cuHandler.createDataTable($(this))}),cuHandler.checkAjaxAction(response),!1):!1},"json"),!1)},submitAjaxForm:function(t){if(!$(t).valid())return!1;cuHandler.showLoader();var e=t.serialize();e+="&CUTOKEN_REQUEST="+$("#cu-token").val();var o=t.attr("action"),i=t.attr("method");return"post"==i?$.post(o,e,cuHandler.retrieveAjax,"json"):$.get(o,e,cuHandler.retrieveAjax,"json"),!1},requestAjax:function(t,e){return $(t).cuSpinner({icon:"svg-rmcommon-spinner-14"}),void 0==e?!1:void 0==e.url?!1:(void 0==e.parameters&&(e.parameters={}),e.parameters.CUTOKEN_REQUEST=$("#cu-token").val(),void(void 0==e.method||"post"==e.method?$.post(e.url,e.parameters,cuHandler.retrieveAjax,"json"):$.get(e.url,e.parameters,cuHandler.retrieveAjax,"json")))},isDisabled:function(t){return $(t).hasClass("disabled")||void 0!=$(t).attr("disabled")?!0:!1},retrieveAjax:function(t){return this.currentResponse=t,"error"==t.type&&(void 0!=t.modal_message?bootbox.alert({message:t.message}):void 0!=t.notify&&void 0!=t.message?(cuHandler.notify({title:void 0==t.notify.title?"":t.notify.title,type:void 0==t.notify.type?"alert-info":t.notify.type,icon:void 0==t.notify.icon?"svg-rmcommon-info-solid":t.notify.icon,text:t.message}),t.notify=void 0):alert(t.message)),void 0!=t.token&&""!=t.token&&$("#cu-token").val(t.token),cuHandler.closeLoader(),cuHandler.checkAjaxAction(t),"error"==t.type?!1:!0},checkAjaxAction:function(data){return void 0!=data.showMessage&&alert(data.message),void 0!=data.notify&&void 0!=data.message&&this.notify({title:void 0==data.notify.title?null:data.notify.title,type:void 0==data.notify.type?"alert-info":data.notify.type,icon:void 0==data.notify.icon?"svg-rmcommon-info-solid":data.notify.icon,text:data.message}),void 0!=data.closeWindow&&$(data.closeWindow).modal("hide"),void 0!=data.runHandler&&eval(data.runHandler+"(data)"),void 0!=data["goto"]&&(window.location.href=data["goto"]),void 0!=data["function"]&&eval(data["function"]),void 0!=data.openDialog&&bootbox.dialog({message:data.content,title:data.message,icon:void 0!=data.icon?data.icon:"",width:void 0!=data.width?data.width:"",owner:void 0!=data.owner?data.owner:"",id:void 0!=data.windowId?data.windowId:"",animate:!1,closeButton:void 0!=data.closeButton?data.closeButton:!0}),void 0!=data.reload?void window.location.reload(!0):void 0},showLoader:function(){$(".cu-window-loader").hide(),$(".cu-window-blocker").hide();var t='<div class="cu-window-blocker"></div>';t+='<div class="cu-window-loader"><div class="loader-container text-center"><button class="close" type="button">&times;</button><span><span class="cu-icon cu-spinner" style="animation: cu-spin 1s infinite steps(10)"><svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14"><path d="M5.233 2.056C5.23 1.063 6.033.253 7.027.25c.994-.003 1.803.8 1.806 1.794.004.994-.8 1.802-1.794 1.806-1 .003-1.81-.8-1.81-1.794zm3.824 1.568c-.003-.993.8-1.802 1.792-1.806.99-.003 1.8.8 1.8 1.793 0 1-.8 1.81-1.8 1.81-1 .01-1.8-.8-1.81-1.79zm2.495 3.81c-.002-.497.398-.9.896-.903.497 0 .902.4.904.9 0 .5-.402.9-.897.91-.498.01-.902-.39-.903-.89zm-1.57 3.824c-.003-.496.4-.9.895-.903.498 0 .903.4.904.897.01.498-.4.9-.89.904-.49.002-.9-.4-.9-.898zM6.17 12.853c-.002-.496.4-.9.897-.902.497 0 .9.4.903.9.002.5-.4.9-.897.91-.498.01-.902-.4-.903-.89zm-3.824-1.57c-.002-.495.4-.9.897-.902.496 0 .9.4.902.9.002.5-.4.9-.896.91-.5 0-.91-.4-.91-.9zM1.87 3.65c-.003-.746.6-1.352 1.345-1.354.746-.002 1.352.6 1.354 1.345 0 .75-.6 1.36-1.35 1.36-.75.01-1.35-.6-1.36-1.34zM.637 7.472c-.002-.56.45-1.014 1.01-1.016.56-.002 1.013.45 1.015 1.01 0 .56-.45 1.014-1.008 1.016S.64 8.03.638 7.472z"/></svg></span> '+cuLanguage.inProgress+"</span></div></div>",$("body").append(t),$(".cu-window-blocker").fadeIn(0,function(){$(".cu-window-loader").fadeIn(1)})},closeLoader:function(t){$(".cu-window-loader").fadeOut(1,function(){$(".cu-window-blocker").fadeOut(0,function(){$(".cu-window-loader").remove(),$(".cu-window-blocker").remove()})})},getURI:function(t,e,o,i,n){return this.getControllerURI(t,e,o,i,n)},getControllerURI:function(module,controller,action,zone,params){var url=xoUrl;if(void 0!=cu_modules[module]&&void 0!=controller&&""!=controller)url+="backend"==zone?"/admin":"",url+=cu_modules[module]+"/"+controller+"/"+action+"/";else{if(url+="/modules/"+module,url+="backend"==zone?"/admin":"",""==controller||void 0==controller)return url;url+="/index.php/"+controller+"/"+action+"/"}if(void 0==params)return url;var query="";for(key in params)query+=(""==query?"?":"&")+key+"="+eval("params."+key);return url+query},url:function(t,e){var o=window.location.protocol+"//"+window.location.host;""!=window.location.port&&(o+=":"+window.location.port);var i=xoUrl.replace(o,"");return void 0==t?xoUrl:1==arguments.length||1!=e?xoUrl+t:i+t},createDataTable:function(t){if(!t.hasClass("dataTable")){var e=$(t).data("exclude");void 0!=e?e.toString().split(","):"";$(t).dataTable({bProcessing:!0,bServerSide:!0,sAjaxSource:$(t).data("source"),bPaginate:!0})}},runAction:function(e){var action=$(e).data("action");switch(action){case"load-remote-dialog":case"load-module-dialog":cuHandler.loadRemoteDialog(e);break;case"goto":var url=$(e).data("url");void 0==url&&(url=$(e).attr("href"));var retriever=$(e).data("retriever");void 0!=retriever&&(url+=eval(retriever+"(e)")),void 0!=url&&""!=url&&(window.location.href=url);break;case"ajax":var url=""!=$(e).attr("href")?$(e).attr("href"):$(e).data("url");cuHandler.requestAjax(e,{url:url})}},enableCommands:function(id_activator,type){var commands=$("*[data-activator='"+id_activator+"']"),total=$("#"+id_activator+" :"+type+"[data-switch]:checked").length;$(commands).each(function(index){var required=void 0!=$(this).data("oncount")?$(this).data("oncount"):" >= 1";eval("total "+required)?$(this).enable():$(this).disable()})},getIcon:function(t){var e="svg-"==t.slice(0,4),o=!1,i=!1;if(e||(e=".svg"==t.slice(-4),e&&(i=!0)),o=null!=t.match(/^(http:\/\/|ftp:\/\/|https:\/\/\|\/\/)/i),e&&!i){var n=t.split("-");if(n.length<3)return this.url("modules/rmcommon/icons/noicon.svg");var s=t.slice(5+n[1].length),a=this.url("modules/rmcommon/icons/"+s+".svg");if(void 0==iconsProviders)return a;if(iconsProviders.hasOwnProperty(n[1])){var r=iconsProviders[n[1]]+"/"+s+".svg";return r}return a}if(i&&!o)return this.url(t);var c=[".jpg",".gif",".png","jpeg"],l=t.slice(-4);return c.indexOf(l)||o?t:'<span class="'+t+'"></span>'},loadIcon:function(t,e,o){if(arguments.length<1)return!1;o=2==arguments.length?!0:1==arguments.length?!1:o;var i=this.getIcon(t),n=".svg"==i.slice(-4),s="<span"==i.slice(0,5);if(o)var a=e.find(".cu-icon");else var a=$("<span />",{"class":"cu-icon"});if((void 0==a||a.length<=0)&&(o=!1,a=$("<span />",{"class":"cu-icon"})),n)a.html("").load(i);else if(s)a.html("").append(i);else{var r=$("<img>").attr("src",i);a.html("").append(r)}return!o&&arguments.length>1&&$(e).append(a),a},modal:bootbox,notify:function(t){return new PNotify(t)}};Number.prototype.formatMoney=function(t,e,o){var i=this,t=isNaN(t=Math.abs(t))?2:t,e=void 0==e?".":e,o=void 0==o?",":o,n=0>i?"-":"",s=parseInt(i=Math.abs(+i||0).toFixed(t))+"",a=(a=s.length)>3?a%3:0;return n+(a?s.substr(0,a)+o:"")+s.substr(a).replace(/(\d{3})(?=\d)/g,"$1"+o)+(t?e+Math.abs(i-s).toFixed(t).slice(2):"")},jQuery.fn.enable=function(){this.each(function(){jQuery(this).removeAttr("disabled"),jQuery(this).removeClass("disabled")})},jQuery.fn.disable=function(){this.each(function(){jQuery(this).attr("disabled",!0),jQuery(this).addClass("disabled")})},$(document).ready(function(){var t="";if($("body").on("click","*[data-action]",function(t){return $(this).is(":disabled")||$(this).attr("disabled")||$(this).hasClass("disabled")?($(t).stopPropagation(),!1):(cuHandler.runAction($(this)),!1)}),$("body").on("submit",'form[data-type="ajax"]',function(){return cuHandler.submitAjaxForm($(this)),!1}),$("body").on("submit",'form[data-type="no-submit"]',function(){return!1}),$("body").on("click",".cu-window-loader .close",function(){cuHandler.closeLoader()}),$(".cu-data-table").each(function(){cuHandler.createDataTable($(this))}),$("*[data-rel='tooltip']").tooltip(),$("body").on("change",".activator-container :checkbox[data-switch], .activator-container :radio[data-switch]",function(){event.stopPropagation();var t=$(this).parents(".activator-container").attr("id");cuHandler.enableCommands(t,$(this).attr("type")),"radio"==$(this).attr("type")&&$(this).parents(".activator-container").find(".tr-checked").removeClass("tr-checked"),$(this).is(":checked")?$(this).parents("tr").addClass("tr-checked"):$(this).parents("tr").removeClass("tr-checked")}),$("body").on("click",".activator-container > tbody > tr",function(t){if("DIV"==t.target.tagName||"TD"==t.target.tagName){var e=$(this).find("input[data-switch]");"radio"==$(e).attr("type")&&$(this).parents(".activator-container").find(".tr-checked").removeClass("tr-checked"),e.is(":checked")?e.removeAttr("checked"):e.prop("checked","checked"),cuHandler.enableCommands($(this).parents(".activator-container").attr("id"),e.attr("type")),$(e).is(":checked")?$(this).addClass("tr-checked"):$(this).removeClass("tr-checked"),event.stopPropagation()}}),$("body").on("change",":checkbox[data-checkbox]",function(){var t=$(this).data("checkbox");if(void 0!=t){$(":checkbox[data-oncheck='"+t+"']").prop("checked",$(this).prop("checked")).change();var e=$(this).parents(".activator-container");e.length>0&&cuHandler.enableCommands(e.attr("id"),"checkbox")}}),$("body").on("change",":checkbox",function(){if(!this.hasAttribute("data-checkbox")&&this.hasAttribute("data-oncheck")){var t=$(":checkbox[data-oncheck='"+$(this).data("oncheck")+"']"),e=$(":checkbox[data-oncheck='"+$(this).data("oncheck")+"']:checked"),o=$(":checkbox[data-checkbox='"+$(this).data("oncheck")+"']");e.length<t.length?$(o).removeAttr("checked"):e.length==t.length&&$(o).prop("checked","checked")}}),1==$("*[data-news='load']").length||$("*[data-boxes='load']").length>0){var e=$("*[data-news='load']");if(e.length<=0&&(e=$("*[data-boxes='load']")),e.length<=0)return!1;var o=e.data("module"),n=$(e.data("target"));void 0!=n&&n.html('<div class="text-success"><span class="fa fa-spinner fa-pulse"></span> '+cuLanguage.downloadNews+"</div>");var s=$("*[data-boxes='load']"),a={module:o,CU_TOKEN:$("#cu-token").val()};$.get(xoUrl+"/modules/rmcommon/ajax/module-info.php",a,function(t){if("error"==t.type)return void n.html('<div class="text-danger"><span class="fa fa-exclamation-triangle"></span> '+cuLanguage.downloadNewsError+"</div>");if(void 0!=t.news&&void 0!=n){for(news=$("<ul>").addClass("cu-ajax-news list-unstyled"),i=0;i<t.news.length;i++){var e="<li><small>"+t.news[i].date+'</small><h5><a href="'+t.news[i].link+'" target="_blank">'+t.news[i].title+"</a></h5>";e+='<p class="help-block">'+t.news[i].content+"</p></li>",news.append(e)}n.html("").append(news),news.fadeIn("fast",function(){$('html.dashboard [data-container="dashboard"]').trigger("containerUpdated")})}if(void 0!=t.boxes&&void 0!=s)for(i=0;i<t.boxes.length;i++){if(void 0==t.boxes[i].size||t.boxes[i].size<=0)var o=1;else o=t.boxes[i].size;var a=$('<div data-dashboard="item">').addClass("size-"+o).css("display","none");if(a.append('<div class="cu-box '+t.boxes[i]["class"]+'"><div class="box-header"><span class="fa fa-caret-up box-handler"></span><h3 class="box-title">'+t.boxes[i].title+'</h3></div><div class="box-content">'+t.boxes[i].content+"</div></div>"),void 0!=t.boxes[i].container){var r=$(t.boxes[i].container);r.length>0&&$(r).each(function(){var e=a.clone();"top"==t.boxes[i].position?$(this).prepend(e):$(this).append(e),e.fadeIn("fast",function(){$('html.dashboard [data-container="dashboard"]').trigger("containerUpdated")})})}}},"json")}$("body").on("click",".ed-container .full-screen",function(){$(this).parents(".ed-container").addClass("full-screen-edit"),$(this).removeClass("full-screen").addClass("normal-screen"),t=$(this).parents(".ed-container").find(".txtarea-container").attr("style"),$(this).parents(".ed-container").find(".txtarea-container").attr("style",""),$("body").css("overflow","hidden")}),$("body").on("click",".ed-container .normal-screen",function(){$(this).parents(".ed-container").removeClass("full-screen-edit"),$(this).addClass("full-screen").removeClass("normal-screen"),$(this).parents(".ed-container").find(".txtarea-container").attr("style",t),t="",$("body").css("overflow","visible")}),$(".cu-notifications form .notification-item").change(function(){var t=$(this).data("event"),e=$(this).is(":checked")?1:0,o=$(this),i={event:t,status:e,page:"cu-notification-subscribe"};$.post(xoUrl+"/notifications.php",i,function(t){return"error"==t.type?void alert(t.message):void $(o).parent().parent().animate({backgroundColor:"#ffff99"},500,"linear",function(){$(this).animate({backgroundColor:"transparent"},500,"linear")})},"json")}),$(" #cu-notifications .cancel-subscription").click(function(){var t=$(this).data("info"),e=0,o=$(this),i={event:t,status:e,page:"cu-notification-subscribe"};$.post(xoUrl+"/notifications.php",i,function(t){return"error"==t.type?void alert(t.message):void $(o).parent().parent().slideUp(250,function(){$(this).remove()})},"json")})});
