+++ src/helium/less/helium/mixins/svg-icons.less
@@ -43,4 +43,24 @@ Format for SVG icons
   line-height: 1;
   vertical-align: text-top;
   vertical-align: middle;
+
+  &.icon2x{
+    .cu-icon-size(32px);
+  }
+
+  &.icon3x{
+    .cu-icon-size(48px);
+  }
+
+  &.icon4x{
+    .cu-icon-size(64px);
+  }
+
+  &.icon5x{
+    .cu-icon-size(96px);
+  }
+
+  &.icon6x{
+    .cu-icon-size(128px);
+  }
 }
